+++ src/device-mapping.ts
@@ -1,21 +1,11 @@
-import Iphone12Mini from './frames/iphone-12-mini';
-import Iphone12ProMax from './frames/iphone-12-pro-max';
-import Iphone14ProMax from './frames/iphone-14-pro-max';
-import Iphone15Pro from './frames/iphone-15';
-import Iphone16Pro from './frames/iphone-16';
+import Iphone15 from './frames/iphone-15';
+import Iphone15Pro from './frames/iphone-15-pro';
+import Iphone16Pro from './frames/iphone-16-pro';
 
 export const DEVICE_MAPPING = {
-  'iphone-12-mini': {
-    component: Iphone12Mini,
-    displayName: 'Iphone 12 Mini',
-  },
-  'iphone-12-pro-max': {
-    component: Iphone12ProMax,
-    displayName: 'Iphone 12 Pro Max',
-  },
-  'iphone-14-pro-max': {
-    component: Iphone14ProMax,
-    displayName: 'Iphone 14 Pro Max',
+  'iphone-15': {
+    component: Iphone15,
+    displayName: 'Iphone 15',
   },
   'iphone-15-pro': {
     component: Iphone15Pro,
