--- src/frames/iphone-12-mini/index.tsx
@@ -1,76 +1,0 @@
-import Wallpaper from '@/assets/images/iphone 12 wallpaper.jpg';
-import AppScreen from "@/components/app-screen";
-import HomeScreen from "@/components/home-screen";
-import Wrapper from "@/components/wrapper";
-import { percentage, pick, px } from "@/lib/utils";
-import { callRef } from "nixix/primitives";
-import { Container } from "nixix/view-components";
-import { containerStyles } from "~/constants";
-import { setupResizeEffect, useIphoneConfig } from "~/stores/iphone-config";
-import DeviceFrame from "./svg/device-frame";
-import StatusBar from "./svg/status-bar";
-
-type Props = App.DeviceProps;
-
-const dimensions = { w: 375, h: 812 };
-
-const deviceWidthRatio = 43 / dimensions.w;
-
-const deviceHeightRatio = 43 / dimensions.h;
-
-const safeAreaInsetRatio = 46 / dimensions.h;
-
-const deviceBarRatios = [
-  18 / dimensions.h,
-  6 / dimensions.h,
-] as const;
-
-const virtualHomeButtonRatio = 130 / dimensions.w;
-
-const clothoidRadiusRatio = 38 / dimensions.w;
-
-const Iphone12Mini: Nixix.FC<Props> = ({ iframeSrc }): someView => {
-  const wrapperRef = callRef<HTMLElement>();
-  const { iphoneConfig } = useIphoneConfig()
-  setupResizeEffect(wrapperRef, {
-    deviceBarRatios,
-    deviceHeightRatio,
-    deviceWidthRatio,
-    clothoidRadiusRatio,
-    virtualHomeButtonRatio,
-    safeAreaInsetRatio,
-  });
-  return (
-    <Wrapper bind:ref={wrapperRef}>
-      <DeviceFrame height={dimensions.h} />
-      <Container
-        style={{
-          ...pick(iphoneConfig, "width", "height"),
-          ...containerStyles,
-        clipPath: iphoneConfig.clothoidRadius,
-        background: `url(${Wallpaper})`,
-        backgroundSize: 'cover'
-      }} >
-        <Container style={{
-          paddingTop: iphoneConfig.safeAreaInset,
-          width: percentage(100),
-          height: percentage(100),
-          clipPath: 'inherit',
-          overflow: 'hidden',
-          position: 'relative'
-        }} >
-          <StatusBar style={{
-            width: iphoneConfig.width,
-            position: 'absolute',
-            top: px(0),
-            zIndex: 900,
-          }} />
-          <HomeScreen iframeSrc={iframeSrc} />
-          <AppScreen config="iphone" iframeSrc={iframeSrc} />
-        </Container>
-      </Container>
-    </Wrapper>
-  );
-};
-
-export default Iphone12Mini;
