--- src/frames/iphone-12-mini/svg/status-bar.tsx
@@ -1,65 +1,0 @@
-
-
-const StatusBar: Nixix.FC<App.SVGProps> = ({ className, ...rest }): someView => {
-  return (
-    <svg
-      {...rest} className={className + ' tws-h-fit '}
-      width="375" height="46" viewBox="0 0 375 46" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <rect width="375" height="46" fill="inherit" />
-      <path d="M36.25 22.3047C36.25 23.2214 36.1771 24.0417 36.0312 24.7656C35.8854 25.4844 35.651 26.0964 35.3281 26.6016C35.0104 27.1068 34.5885 27.4922 34.0625 27.7578C33.5365 28.0234 32.8906 28.1562 32.125 28.1562C31.1719 28.1562 30.3906 27.9245 29.7812 27.4609C29.1719 26.9922 28.7214 26.3229 28.4297 25.4531C28.1432 24.5781 28 23.5286 28 22.3047C28 21.0703 28.1302 20.0156 28.3906 19.1406C28.6562 18.2604 29.0911 17.5859 29.6953 17.1172C30.2995 16.6484 31.1094 16.4141 32.125 16.4141C33.0781 16.4141 33.8568 16.6484 34.4609 17.1172C35.0703 17.5807 35.5208 18.2526 35.8125 19.1328C36.1042 20.0078 36.25 21.0651 36.25 22.3047ZM31.0703 22.3047C31.0703 23.0495 31.0964 23.6745 31.1484 24.1797C31.2057 24.6849 31.3099 25.0651 31.4609 25.3203C31.6172 25.5703 31.8385 25.6953 32.125 25.6953C32.4115 25.6953 32.6302 25.5703 32.7812 25.3203C32.9323 25.0651 33.0365 24.6875 33.0938 24.1875C33.151 23.6823 33.1797 23.0547 33.1797 22.3047C33.1797 21.5495 33.151 20.9193 33.0938 20.4141C33.0365 19.9089 32.9323 19.5286 32.7812 19.2734C32.6302 19.013 32.4115 18.8828 32.125 18.8828C31.8385 18.8828 31.6172 19.013 31.4609 19.2734C31.3099 19.5286 31.2057 19.9089 31.1484 20.4141C31.0964 20.9193 31.0703 21.5495 31.0703 22.3047ZM45.6406 21.5859C45.6406 22.4036 45.5677 23.1536 45.4219 23.8359C45.2812 24.5182 45.0599 25.1276 44.7578 25.6641C44.4609 26.1953 44.0781 26.6484 43.6094 27.0234C43.1406 27.3932 42.5807 27.6745 41.9297 27.8672C41.2786 28.0599 40.5286 28.1562 39.6797 28.1562C39.4453 28.1562 39.1719 28.151 38.8594 28.1406C38.5521 28.1302 38.2917 28.1094 38.0781 28.0781V25.6328C38.3021 25.6693 38.526 25.7005 38.75 25.7266C38.974 25.7474 39.2344 25.7578 39.5312 25.7578C40.3906 25.7578 41.0547 25.6354 41.5234 25.3906C41.9974 25.1458 42.3307 24.8073 42.5234 24.375C42.7161 23.9375 42.8281 23.4349 42.8594 22.8672H42.7656C42.6302 23.1224 42.4661 23.3516 42.2734 23.5547C42.0807 23.7526 41.8385 23.9089 41.5469 24.0234C41.2552 24.138 40.8932 24.1953 40.4609 24.1953C39.8203 24.1953 39.263 24.0495 38.7891 23.7578C38.3203 23.4661 37.9557 23.0443 37.6953 22.4922C37.4349 21.9349 37.3047 21.2578 37.3047 20.4609C37.3047 19.612 37.4661 18.888 37.7891 18.2891C38.1172 17.6849 38.5833 17.2214 39.1875 16.8984C39.7969 16.5755 40.5208 16.4141 41.3594 16.4141C41.9531 16.4141 42.5078 16.5156 43.0234 16.7188C43.5443 16.9219 44 17.2344 44.3906 17.6562C44.7812 18.0729 45.0859 18.6068 45.3047 19.2578C45.5286 19.9089 45.6406 20.6849 45.6406 21.5859ZM41.4375 18.8828C41.0781 18.8828 40.7969 19.0052 40.5938 19.25C40.3958 19.4948 40.2969 19.888 40.2969 20.4297C40.2969 20.8568 40.3906 21.1953 40.5781 21.4453C40.7656 21.6953 41.0417 21.8203 41.4062 21.8203C41.6354 21.8203 41.8359 21.7604 42.0078 21.6406C42.1849 21.5156 42.3229 21.3594 42.4219 21.1719C42.5208 20.9792 42.5703 20.7812 42.5703 20.5781C42.5703 20.375 42.5469 20.1745 42.5 19.9766C42.4583 19.7734 42.3906 19.5911 42.2969 19.4297C42.2083 19.263 42.0911 19.1302 41.9453 19.0312C41.8047 18.9323 41.6354 18.8828 41.4375 18.8828ZM46.8594 26.7031C46.8594 26.151 47.0208 25.763 47.3438 25.5391C47.6719 25.3151 48.0651 25.2031 48.5234 25.2031C48.9557 25.2031 49.3307 25.3151 49.6484 25.5391C49.9661 25.763 50.125 26.151 50.125 26.7031C50.125 27.2292 49.9661 27.6094 49.6484 27.8438C49.3307 28.0781 48.9557 28.1953 48.5234 28.1953C48.0651 28.1953 47.6719 28.0781 47.3438 27.8438C47.0208 27.6094 46.8594 27.2292 46.8594 26.7031ZM46.8594 20.5312C46.8594 19.9844 47.0208 19.599 47.3438 19.375C47.6719 19.151 48.0651 19.0391 48.5234 19.0391C48.9557 19.0391 49.3307 19.151 49.6484 19.375C49.9661 19.599 50.125 19.9844 50.125 20.5312C50.125 21.0677 49.9661 21.4531 49.6484 21.6875C49.3307 21.9167 48.9557 22.0312 48.5234 22.0312C48.0651 22.0312 47.6719 21.9167 47.3438 21.6875C47.0208 21.4531 46.8594 21.0677 46.8594 20.5312ZM59.9453 25.7891H58.6875V28H55.6406V25.7891H51.0859V23.5234L55.8359 16.5781H58.6875V23.4844H59.9453V25.7891ZM55.6406 23.4844V22.1562C55.6406 22.0365 55.6432 21.8854 55.6484 21.7031C55.6536 21.5208 55.6589 21.3385 55.6641 21.1562C55.6745 20.9688 55.6823 20.8125 55.6875 20.6875C55.6979 20.5573 55.7057 20.487 55.7109 20.4766H55.625C55.526 20.7109 55.4349 20.9062 55.3516 21.0625C55.2682 21.2135 55.1589 21.3932 55.0234 21.6016L53.7578 23.4844H55.6406ZM67.5469 28H64.4141V22.3516C64.4141 22.2474 64.4141 22.0573 64.4141 21.7812C64.4141 21.5 64.4167 21.1901 64.4219 20.8516C64.4323 20.513 64.4453 20.2083 64.4609 19.9375C64.3255 20.099 64.1979 20.237 64.0781 20.3516C63.9635 20.4661 63.8516 20.5729 63.7422 20.6719L62.4453 21.7422L60.8438 19.7734L64.7656 16.5781H67.5469V28Z" fill="#111111" />
-      <g clip-path="url(#clip0_311_422)">
-        <path d="M92 0H283V0C283 9.3104 283 13.9656 281.672 17.7022C279.354 24.2234 274.223 29.3544 267.702 31.672C263.966 33 259.31 33 250 33H125C115.69 33 111.034 33 107.298 31.672C100.777 29.3544 95.6456 24.2234 93.328 17.7022C92 13.9656 92 9.3104 92 0V0Z" fill="black" />
-        <rect x="159" y="12" width="48" height="8" rx="4" fill="#171717" fill-opacity="0.7" />
-        <path d="M223.985 22.1965C227.526 22.1965 230.396 19.4662 230.396 16.0982C230.396 12.7303 227.526 10 223.985 10C220.445 10 217.575 12.7303 217.575 16.0982C217.575 19.4662 220.445 22.1965 223.985 22.1965Z" fill="#161616" />
-        <path d="M223.986 19.859C226.17 19.859 227.94 18.175 227.94 16.0976C227.94 14.0202 226.17 12.3361 223.986 12.3361C221.802 12.3361 220.032 14.0202 220.032 16.0976C220.032 18.175 221.802 19.859 223.986 19.859Z" fill="#0A0D13" />
-        <path d="M223.985 18.9636C225.649 18.9636 226.998 17.6805 226.998 16.0977C226.998 14.5149 225.649 13.2317 223.985 13.2317C222.322 13.2317 220.973 14.5149 220.973 16.0977C220.973 17.6805 222.322 18.9636 223.985 18.9636Z" fill="#091427" />
-        <g filter="url(#filter0_f_311_422)">
-          <path fill-rule="evenodd" clip-rule="evenodd" d="M221.912 14.1805C221.555 14.2163 221.047 14.735 221.009 15.7546C220.972 16.7741 221.442 17.3286 221.705 17.3286C221.968 17.3286 223.059 15.844 221.912 14.1805Z" fill="#235A91" fill-opacity="0.556075" />
-        </g>
-        <g filter="url(#filter1_f_311_422)">
-          <path fill-rule="evenodd" clip-rule="evenodd" d="M224.477 13.8237C223.914 13.9886 223.806 14.8546 223.827 15.3494C223.849 15.8443 224.586 17.1432 225.475 16.8546C226.363 16.5659 226.515 15.4938 226.146 14.7721C225.778 14.0505 225.106 13.535 224.477 13.8237Z" fill="#235A91" fill-opacity="0.556075" />
-        </g>
-      </g>
-      <g clip-path="url(#clip1_311_422)">
-        <g clip-path="url(#clip2_311_422)">
-          <path fill-rule="evenodd" clip-rule="evenodd" d="M307.149 20C306.597 20 306.149 20.4477 306.149 21V26.2C306.149 26.7523 306.597 27.2 307.149 27.2H308.5C309.052 27.2 309.5 26.7523 309.5 26.2V21C309.5 20.4477 309.052 20 308.5 20H307.149ZM301.6 21.3C301.047 21.3 300.6 21.7478 300.6 22.3001V26.2C300.6 26.7523 301.047 27.2 301.6 27.2H302.95C303.503 27.2 303.95 26.7523 303.95 26.2V22.3C303.95 21.7478 303.503 21.3 302.95 21.3H301.6ZM295.05 23.6C295.05 23.0477 295.498 22.6 296.05 22.6H297.401C297.953 22.6 298.401 23.0477 298.401 23.6V26.2C298.401 26.7523 297.953 27.2 297.401 27.2H296.05C295.498 27.2 295.05 26.7523 295.05 26.2V23.6ZM290.5 23.9C289.948 23.9 289.5 24.3477 289.5 24.9V26.2C289.5 26.7523 289.948 27.2 290.5 27.2H291.851C292.403 27.2 292.851 26.7523 292.851 26.2V24.9C292.851 24.3477 292.403 23.9 291.851 23.9H290.5ZM306.149 29.8C306.149 29.2478 306.597 28.8 307.149 28.8H308.5C309.052 28.8 309.5 29.2478 309.5 29.8V31C309.5 31.5523 309.052 32 308.5 32H307.149C306.597 32 306.149 31.5523 306.149 31V29.8ZM301.6 28.8C301.047 28.8 300.6 29.2478 300.6 29.8V31C300.6 31.5523 301.047 32 301.6 32H302.95C303.503 32 303.95 31.5523 303.95 31V29.8C303.95 29.2478 303.503 28.8 302.95 28.8H301.6ZM295.05 29.8C295.05 29.2478 295.498 28.8 296.05 28.8H297.401C297.953 28.8 298.401 29.2478 298.401 29.8V31C298.401 31.5523 297.953 32 297.401 32H296.05C295.498 32 295.05 31.5523 295.05 31V29.8ZM290.5 28.8C289.948 28.8 289.5 29.2478 289.5 29.8V31C289.5 31.5523 289.948 32 290.5 32H291.851C292.403 32 292.851 31.5523 292.851 31V29.8C292.851 29.2478 292.403 28.8 291.851 28.8H290.5Z" fill="#111111" />
-        </g>
-        <path d="M323.509 28.3275C324.227 28.3275 324.925 28.5292 325.539 28.9137L325.788 29.0695C325.977 29.1876 326.011 29.464 325.856 29.629L323.729 31.8956C323.598 32.0348 323.387 32.0348 323.256 31.8956L321.143 29.6436C320.989 29.4793 321.021 29.2043 321.208 29.0853L321.454 28.9288C322.075 28.5346 322.781 28.3275 323.509 28.3275Z" fill="#111111" />
-        <path d="M323.509 24.1639C325.295 24.1639 327.01 24.7972 328.407 25.97L328.604 26.1357C328.762 26.2678 328.776 26.5168 328.634 26.6678L327.364 28.021C327.247 28.1462 327.061 28.1605 326.928 28.0548L326.773 27.9318C325.83 27.1823 324.692 26.7803 323.509 26.7803C322.318 26.7803 321.174 27.1873 320.227 27.9454L320.072 28.0695C319.939 28.1761 319.753 28.1621 319.635 28.0366L318.366 26.6839C318.224 26.5332 318.238 26.2848 318.394 26.1525L318.591 25.9866C319.991 24.8033 321.714 24.1639 323.509 24.1639Z" fill="#111111" />
-        <path d="M323.509 20C326.343 20 329.054 21.0687 331.2 23.0275L331.383 23.1947C331.531 23.3299 331.54 23.5707 331.402 23.7176L330.137 25.0663C330.014 25.197 329.818 25.2062 329.685 25.0874L329.528 24.9474C327.839 23.4377 325.722 22.6164 323.509 22.6164C321.288 22.6164 319.164 23.4433 317.472 24.9625L317.315 25.1033C317.183 25.2225 316.986 25.2136 316.863 25.0827L315.598 23.7341C315.46 23.5874 315.469 23.347 315.616 23.2118L315.799 23.0445C317.947 21.0751 320.666 20 323.509 20Z" fill="#111111" />
-        <g clip-path="url(#clip3_311_422)">
-          <ellipse cx="361" cy="26" rx="2.5" ry="3" fill="#171717" />
-          <rect x="337.5" y="19" width="24" height="14" rx="4" fill="#171717" />
-          <path d="M345.311 30H343.353V26.4697C343.353 26.4046 343.353 26.2858 343.353 26.1133C343.353 25.9375 343.354 25.7438 343.357 25.5322C343.364 25.3206 343.372 25.1302 343.382 24.9609C343.297 25.0618 343.217 25.1481 343.143 25.2197C343.071 25.2913 343.001 25.3581 342.933 25.4199L342.122 26.0889L341.121 24.8584L343.572 22.8613H345.311V30ZM352.078 26.4404C352.078 27.0133 352.033 27.526 351.941 27.9785C351.85 28.4277 351.704 28.8102 351.502 29.126C351.303 29.4417 351.04 29.6826 350.711 29.8486C350.382 30.0146 349.979 30.0977 349.5 30.0977C348.904 30.0977 348.416 29.9528 348.035 29.6631C347.654 29.3701 347.373 28.9518 347.19 28.4082C347.011 27.8613 346.922 27.2054 346.922 26.4404C346.922 25.6689 347.003 25.0098 347.166 24.4629C347.332 23.9128 347.604 23.4912 347.981 23.1982C348.359 22.9053 348.865 22.7588 349.5 22.7588C350.096 22.7588 350.582 22.9053 350.96 23.1982C351.341 23.488 351.622 23.9079 351.805 24.458C351.987 25.0049 352.078 25.6657 352.078 26.4404ZM348.841 26.4404C348.841 26.9059 348.857 27.2965 348.89 27.6123C348.925 27.9281 348.991 28.1657 349.085 28.3252C349.183 28.4814 349.321 28.5596 349.5 28.5596C349.679 28.5596 349.816 28.4814 349.91 28.3252C350.005 28.1657 350.07 27.9297 350.105 27.6172C350.141 27.3014 350.159 26.9092 350.159 26.4404C350.159 25.9684 350.141 25.5745 350.105 25.2588C350.07 24.943 350.005 24.7054 349.91 24.5459C349.816 24.3831 349.679 24.3018 349.5 24.3018C349.321 24.3018 349.183 24.3831 349.085 24.5459C348.991 24.7054 348.925 24.943 348.89 25.2588C348.857 25.5745 348.841 25.9684 348.841 26.4404ZM357.938 26.4404C357.938 27.0133 357.892 27.526 357.801 27.9785C357.71 28.4277 357.563 28.8102 357.361 29.126C357.163 29.4417 356.899 29.6826 356.57 29.8486C356.242 30.0146 355.838 30.0977 355.359 30.0977C354.764 30.0977 354.275 29.9528 353.895 29.6631C353.514 29.3701 353.232 28.9518 353.05 28.4082C352.871 27.8613 352.781 27.2054 352.781 26.4404C352.781 25.6689 352.863 25.0098 353.025 24.4629C353.191 23.9128 353.463 23.4912 353.841 23.1982C354.218 22.9053 354.725 22.7588 355.359 22.7588C355.955 22.7588 356.442 22.9053 356.819 23.1982C357.2 23.488 357.482 23.9079 357.664 24.458C357.846 25.0049 357.938 25.6657 357.938 26.4404ZM354.7 26.4404C354.7 26.9059 354.716 27.2965 354.749 27.6123C354.785 27.9281 354.85 28.1657 354.944 28.3252C355.042 28.4814 355.18 28.5596 355.359 28.5596C355.538 28.5596 355.675 28.4814 355.77 28.3252C355.864 28.1657 355.929 27.9297 355.965 27.6172C356.001 27.3014 356.019 26.9092 356.019 26.4404C356.019 25.9684 356.001 25.5745 355.965 25.2588C355.929 24.943 355.864 24.7054 355.77 24.5459C355.675 24.3831 355.538 24.3018 355.359 24.3018C355.18 24.3018 355.042 24.3831 354.944 24.5459C354.85 24.7054 354.785 24.943 354.749 25.2588C354.716 25.5745 354.7 25.9684 354.7 26.4404Z" fill="#F5F5F5" />
-        </g>
-      </g>
-      <defs>
-        <filter id="filter0_f_311_422" x="220.804" y="13.9777" width="1.82514" height="3.5538" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="0.101426" result="effect1_foregroundBlur_311_422" />
-        </filter>
-        <filter id="filter1_f_311_422" x="223.622" y="13.5392" width="2.91449" height="3.55954" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="0.101426" result="effect1_foregroundBlur_311_422" />
-        </filter>
-        <clipPath id="clip0_311_422">
-          <rect width="191" height="33" fill="white" transform="translate(92)" />
-        </clipPath>
-        <clipPath id="clip1_311_422">
-          <rect width="74" height="14" fill="white" transform="translate(289.5 19)" />
-        </clipPath>
-        <clipPath id="clip2_311_422">
-          <rect width="20" height="14" fill="white" transform="translate(289.5 19)" />
-        </clipPath>
-        <clipPath id="clip3_311_422">
-          <rect width="26" height="14" fill="white" transform="translate(337.5 19)" />
-        </clipPath>
-      </defs>
-    </svg>
-
-  )
-}
-
-export default StatusBar;
