--- src/frames/iphone-12-pro-max/svg/status-bar.tsx
@@ -1,73 +1,0 @@
-
-const StatusBar: Nixix.FC<App.SVGProps> = ({ className, ...rest }): someView => {
-  return (
-    <svg
-      {...rest} className={className + ' tws-h-fit '}
-      width="428" height="55" viewBox="0 0 428 55" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clip-path="url(#clip0_311_518)">
-        <path d="M43.7031 27.7368C43.7031 28.8254 43.6165 29.7995 43.4434 30.6592C43.2702 31.5127 42.9919 32.2394 42.6084 32.8394C42.2311 33.4393 41.7301 33.897 41.1055 34.2124C40.4808 34.5278 39.7139 34.6855 38.8047 34.6855C37.6729 34.6855 36.7451 34.4103 36.0215 33.8599C35.2979 33.3032 34.7629 32.5085 34.4165 31.4756C34.0763 30.4365 33.9062 29.1903 33.9062 27.7368C33.9062 26.271 34.0609 25.0186 34.3701 23.9795C34.6855 22.9342 35.202 22.1333 35.9194 21.5767C36.6369 21.02 37.5986 20.7417 38.8047 20.7417C39.9365 20.7417 40.8612 21.02 41.5786 21.5767C42.3022 22.1271 42.8372 22.925 43.1836 23.9702C43.5299 25.0093 43.7031 26.2648 43.7031 27.7368ZM37.5522 27.7368C37.5522 28.6213 37.5832 29.3634 37.645 29.9634C37.7131 30.5633 37.8368 31.0148 38.0161 31.3179C38.2017 31.6147 38.4645 31.7632 38.8047 31.7632C39.1449 31.7632 39.4046 31.6147 39.584 31.3179C39.7633 31.0148 39.887 30.5664 39.9551 29.9727C40.0231 29.3727 40.0571 28.6274 40.0571 27.7368C40.0571 26.84 40.0231 26.0916 39.9551 25.4917C39.887 24.8918 39.7633 24.4403 39.584 24.1372C39.4046 23.828 39.1449 23.6733 38.8047 23.6733C38.4645 23.6733 38.2017 23.828 38.0161 24.1372C37.8368 24.4403 37.7131 24.8918 37.645 25.4917C37.5832 26.0916 37.5522 26.84 37.5522 27.7368ZM54.8545 26.8833C54.8545 27.8543 54.7679 28.745 54.5947 29.5552C54.4277 30.3654 54.1649 31.089 53.8062 31.7261C53.4536 32.3569 52.999 32.895 52.4424 33.3403C51.8857 33.7795 51.2209 34.1134 50.4478 34.3423C49.6746 34.5711 48.784 34.6855 47.7759 34.6855C47.4976 34.6855 47.1729 34.6794 46.8018 34.667C46.4368 34.6546 46.1276 34.6299 45.874 34.5928V31.689C46.14 31.7323 46.4059 31.7694 46.6719 31.8003C46.9378 31.825 47.2471 31.8374 47.5996 31.8374C48.6201 31.8374 49.4087 31.6921 49.9653 31.4014C50.5282 31.1107 50.924 30.7087 51.1528 30.1953C51.3817 29.6758 51.5146 29.0789 51.5518 28.4048H51.4404C51.2796 28.7078 51.0848 28.98 50.856 29.2212C50.6271 29.4562 50.3395 29.6418 49.9932 29.7778C49.6468 29.9139 49.217 29.9819 48.7036 29.9819C47.9429 29.9819 47.2811 29.8088 46.7183 29.4624C46.1616 29.116 45.7287 28.6151 45.4194 27.9595C45.1102 27.2977 44.9556 26.4937 44.9556 25.5474C44.9556 24.5392 45.1473 23.6795 45.5308 22.9683C45.9204 22.2508 46.474 21.7004 47.1914 21.3169C47.915 20.9334 48.7747 20.7417 49.7705 20.7417C50.4756 20.7417 51.1343 20.8623 51.7466 21.1035C52.3651 21.3447 52.9062 21.7158 53.3701 22.2168C53.834 22.7116 54.1958 23.3455 54.4556 24.1187C54.7215 24.8918 54.8545 25.8133 54.8545 26.8833ZM49.8633 23.6733C49.4365 23.6733 49.1025 23.8187 48.8613 24.1094C48.6263 24.4001 48.5088 24.867 48.5088 25.5103C48.5088 26.0174 48.6201 26.4194 48.8428 26.7163C49.0654 27.0132 49.3932 27.1616 49.8262 27.1616C50.0983 27.1616 50.3364 27.0905 50.5405 26.9482C50.7508 26.7998 50.9147 26.6143 51.0322 26.3916C51.1497 26.1628 51.2085 25.9277 51.2085 25.6865C51.2085 25.4453 51.1807 25.2072 51.125 24.9722C51.0755 24.731 50.9951 24.5145 50.8838 24.3228C50.7786 24.1248 50.6395 23.9671 50.4663 23.8496C50.2993 23.7321 50.0983 23.6733 49.8633 23.6733ZM56.3018 32.96C56.3018 32.3044 56.4935 31.8436 56.877 31.5776C57.2666 31.3117 57.7336 31.1787 58.2778 31.1787C58.7912 31.1787 59.2365 31.3117 59.6138 31.5776C59.991 31.8436 60.1797 32.3044 60.1797 32.96C60.1797 33.5846 59.991 34.0361 59.6138 34.3145C59.2365 34.5928 58.7912 34.7319 58.2778 34.7319C57.7336 34.7319 57.2666 34.5928 56.877 34.3145C56.4935 34.0361 56.3018 33.5846 56.3018 32.96ZM56.3018 25.6309C56.3018 24.9814 56.4935 24.5238 56.877 24.2578C57.2666 23.9919 57.7336 23.8589 58.2778 23.8589C58.7912 23.8589 59.2365 23.9919 59.6138 24.2578C59.991 24.5238 60.1797 24.9814 60.1797 25.6309C60.1797 26.2679 59.991 26.7256 59.6138 27.0039C59.2365 27.276 58.7912 27.4121 58.2778 27.4121C57.7336 27.4121 57.2666 27.276 56.877 27.0039C56.4935 26.7256 56.3018 26.2679 56.3018 25.6309ZM71.8413 31.8745H70.3477V34.5H66.7295V31.8745H61.3208V29.1841L66.9614 20.9365H70.3477V29.1377H71.8413V31.8745ZM66.7295 29.1377V27.5605C66.7295 27.4183 66.7326 27.2389 66.7388 27.0225C66.745 26.806 66.7511 26.5895 66.7573 26.373C66.7697 26.1504 66.779 25.9648 66.7852 25.8164C66.7975 25.6618 66.8068 25.5783 66.813 25.5659H66.7109C66.5934 25.8442 66.4852 26.0762 66.3862 26.2617C66.2873 26.4411 66.1574 26.6545 65.9966 26.9019L64.4937 29.1377H66.7295ZM80.8682 34.5H77.1479V27.7925C77.1479 27.6688 77.1479 27.443 77.1479 27.1152C77.1479 26.7812 77.151 26.4132 77.1572 26.0112C77.1696 25.6092 77.1851 25.2474 77.2036 24.9258C77.0428 25.1175 76.8913 25.2814 76.749 25.4175C76.613 25.5535 76.48 25.6803 76.3501 25.7979L74.8101 27.0688L72.9082 24.731L77.5654 20.9365H80.8682V34.5Z" fill="#111111" />
-        <g clip-path="url(#clip1_311_518)">
-          <g clip-path="url(#clip2_311_518)">
-            <path d="M108 0H319V0C319 9.3104 319 13.9656 317.672 17.7022C315.354 24.2234 310.223 29.3544 303.702 31.672C299.966 33 295.31 33 286 33H141C131.69 33 127.034 33 123.298 31.672C116.777 29.3544 111.646 24.2234 109.328 17.7022C108 13.9656 108 9.3104 108 0V0Z" fill="black" />
-            <rect x="190" y="13" width="48" height="8" rx="4" fill="#171717" fill-opacity="0.7" />
-            <path d="M250.98 22.1965C254.521 22.1965 257.391 19.4662 257.391 16.0982C257.391 12.7303 254.521 10 250.98 10C247.44 10 244.57 12.7303 244.57 16.0982C244.57 19.4662 247.44 22.1965 250.98 22.1965Z" fill="#161616" />
-            <path d="M250.981 19.859C253.165 19.859 254.935 18.1749 254.935 16.0975C254.935 14.0201 253.165 12.3361 250.981 12.3361C248.797 12.3361 247.027 14.0201 247.027 16.0975C247.027 18.1749 248.797 19.859 250.981 19.859Z" fill="#0A0D13" />
-            <path d="M250.98 18.9636C252.644 18.9636 253.993 17.6805 253.993 16.0977C253.993 14.5149 252.644 13.2317 250.98 13.2317C249.316 13.2317 247.967 14.5149 247.967 16.0977C247.967 17.6805 249.316 18.9636 250.98 18.9636Z" fill="#091427" />
-            <g filter="url(#filter0_f_311_518)">
-              <path fill-rule="evenodd" clip-rule="evenodd" d="M248.907 14.1805C248.55 14.2163 248.042 14.735 248.005 15.7546C247.967 16.7741 248.437 17.3286 248.7 17.3286C248.964 17.3286 250.054 15.844 248.907 14.1805Z" fill="#235A91" fill-opacity="0.556075" />
-            </g>
-            <g filter="url(#filter1_f_311_518)">
-              <path fill-rule="evenodd" clip-rule="evenodd" d="M251.473 13.8237C250.909 13.9886 250.801 14.8546 250.822 15.3494C250.844 15.8443 251.581 17.1432 252.47 16.8546C253.358 16.5659 253.51 15.4938 253.142 14.7721C252.773 14.0505 252.101 13.535 251.473 13.8237Z" fill="#235A91" fill-opacity="0.556075" />
-            </g>
-          </g>
-        </g>
-        <g clip-path="url(#clip3_311_518)">
-          <g clip-path="url(#clip4_311_518)">
-            <path fill-rule="evenodd" clip-rule="evenodd" d="M348.063 18.1429C347.511 18.1429 347.063 18.5906 347.063 19.1429V25.3714C347.063 25.9237 347.511 26.3714 348.063 26.3714H349.893C350.445 26.3714 350.893 25.9237 350.893 25.3714V19.1429C350.893 18.5906 350.445 18.1429 349.893 18.1429H348.063ZM341.721 19.6286C341.169 19.6286 340.721 20.0763 340.721 20.6286V25.3715C340.721 25.9238 341.169 26.3715 341.721 26.3715H343.55C344.103 26.3715 344.55 25.9238 344.55 25.3715V20.6286C344.55 20.0763 344.103 19.6286 343.55 19.6286H341.721ZM334.378 22.1143C334.378 21.562 334.826 21.1143 335.378 21.1143H337.208C337.76 21.1143 338.208 21.562 338.208 22.1143V25.3714C338.208 25.9237 337.76 26.3714 337.208 26.3714H335.378C334.826 26.3714 334.378 25.9237 334.378 25.3714V22.1143ZM329.036 22.6C328.483 22.6 328.036 23.0477 328.036 23.6V25.3715C328.036 25.9237 328.483 26.3715 329.036 26.3715H330.865C331.417 26.3715 331.865 25.9237 331.865 25.3715V23.6C331.865 23.0477 331.417 22.6 330.865 22.6H329.036ZM347.063 29.2001C347.063 28.6478 347.511 28.2001 348.063 28.2001H349.893C350.445 28.2001 350.893 28.6478 350.893 29.2001V30.8572C350.893 31.4095 350.445 31.8572 349.893 31.8572H348.063C347.511 31.8572 347.063 31.4095 347.063 30.8572V29.2001ZM341.721 28.2001C341.169 28.2001 340.721 28.6478 340.721 29.2001V30.8572C340.721 31.4095 341.169 31.8572 341.721 31.8572H343.55C344.103 31.8572 344.55 31.4095 344.55 30.8572V29.2001C344.55 28.6478 344.103 28.2001 343.55 28.2001H341.721ZM334.378 29.2001C334.378 28.6478 334.826 28.2001 335.378 28.2001H337.208C337.76 28.2001 338.208 28.6478 338.208 29.2001V30.8572C338.208 31.4095 337.76 31.8572 337.208 31.8572H335.378C334.826 31.8572 334.378 31.4095 334.378 30.8572V29.2001ZM329.036 28.2001C328.483 28.2001 328.036 28.6478 328.036 29.2001V30.8572C328.036 31.4095 328.483 31.8572 329.036 31.8572H330.865C331.417 31.8572 331.865 31.4095 331.865 30.8572V29.2001C331.865 28.6478 331.417 28.2001 330.865 28.2001H329.036Z" fill="#111111" />
-          </g>
-          <path d="M366.332 27.66C367.152 27.66 367.95 27.8905 368.652 28.3299L368.937 28.5079C369.153 28.6429 369.191 28.9588 369.014 29.1474L366.583 31.7378C366.434 31.8969 366.192 31.8969 366.043 31.7378L363.628 29.1641C363.451 28.9763 363.488 28.662 363.702 28.526L363.984 28.3472C364.693 27.8966 365.5 27.66 366.332 27.66Z" fill="#111111" />
-          <path d="M366.332 22.9016C368.373 22.9016 370.333 23.6253 371.93 24.9657L372.155 25.1551C372.335 25.3061 372.351 25.5907 372.189 25.7633L370.738 27.3097C370.604 27.4528 370.391 27.4692 370.239 27.3483L370.062 27.2078C368.984 26.3513 367.684 25.8918 366.332 25.8918C364.971 25.8918 363.663 26.357 362.581 27.2234L362.404 27.3651C362.252 27.487 362.039 27.471 361.904 27.3275L360.454 25.7816C360.292 25.6094 360.308 25.3255 360.487 25.1742L360.711 24.9847C362.311 23.6324 364.28 22.9016 366.332 22.9016Z" fill="#111111" />
-          <path d="M366.331 18.1429C369.571 18.1429 372.669 19.3642 375.122 21.6029L375.331 21.794C375.5 21.9484 375.51 22.2236 375.352 22.3915L373.906 23.933C373.766 24.0823 373.542 24.0928 373.39 23.9571L373.211 23.797C371.28 22.0716 368.86 21.1331 366.331 21.1331C363.794 21.1331 361.366 22.0781 359.433 23.8143L359.253 23.9751C359.102 24.1115 358.877 24.1012 358.737 23.9516L357.291 22.4104C357.133 22.2428 357.143 21.968 357.312 21.8134L357.52 21.6223C359.975 19.3715 363.082 18.1429 366.331 18.1429Z" fill="#111111" />
-          <g clip-path="url(#clip5_311_518)">
-            <ellipse cx="408.607" cy="25" rx="2.85714" ry="3.42858" fill="#171717" />
-            <rect x="381.75" y="17" width="27.4287" height="16" rx="4" fill="#171717" />
-            <path d="M391.274 29H389.316V25.4697C389.316 25.4046 389.316 25.2858 389.316 25.1133C389.316 24.9375 389.318 24.7438 389.321 24.5322C389.328 24.3206 389.336 24.1302 389.346 23.9609C389.261 24.0618 389.181 24.1481 389.106 24.2197C389.035 24.2913 388.965 24.3581 388.896 24.4199L388.086 25.0889L387.085 23.8584L389.536 21.8613H391.274V29ZM398.042 25.4404C398.042 26.0133 397.996 26.526 397.905 26.9785C397.814 27.4277 397.668 27.8102 397.466 28.126C397.267 28.4417 397.004 28.6826 396.675 28.8486C396.346 29.0146 395.942 29.0977 395.464 29.0977C394.868 29.0977 394.38 28.9528 393.999 28.6631C393.618 28.3701 393.337 27.9518 393.154 27.4082C392.975 26.8613 392.886 26.2054 392.886 25.4404C392.886 24.6689 392.967 24.0098 393.13 23.4629C393.296 22.9128 393.568 22.4912 393.945 22.1982C394.323 21.9053 394.829 21.7588 395.464 21.7588C396.06 21.7588 396.546 21.9053 396.924 22.1982C397.305 22.488 397.586 22.9079 397.769 23.458C397.951 24.0049 398.042 24.6657 398.042 25.4404ZM394.805 25.4404C394.805 25.9059 394.821 26.2965 394.854 26.6123C394.889 26.9281 394.954 27.1657 395.049 27.3252C395.146 27.4814 395.285 27.5596 395.464 27.5596C395.643 27.5596 395.78 27.4814 395.874 27.3252C395.968 27.1657 396.034 26.9297 396.069 26.6172C396.105 26.3014 396.123 25.9092 396.123 25.4404C396.123 24.9684 396.105 24.5745 396.069 24.2588C396.034 23.943 395.968 23.7054 395.874 23.5459C395.78 23.3831 395.643 23.3018 395.464 23.3018C395.285 23.3018 395.146 23.3831 395.049 23.5459C394.954 23.7054 394.889 23.943 394.854 24.2588C394.821 24.5745 394.805 24.9684 394.805 25.4404ZM403.901 25.4404C403.901 26.0133 403.856 26.526 403.765 26.9785C403.674 27.4277 403.527 27.8102 403.325 28.126C403.127 28.4417 402.863 28.6826 402.534 28.8486C402.205 29.0146 401.802 29.0977 401.323 29.0977C400.728 29.0977 400.239 28.9528 399.858 28.6631C399.478 28.3701 399.196 27.9518 399.014 27.4082C398.835 26.8613 398.745 26.2054 398.745 25.4404C398.745 24.6689 398.826 24.0098 398.989 23.4629C399.155 22.9128 399.427 22.4912 399.805 22.1982C400.182 21.9053 400.688 21.7588 401.323 21.7588C401.919 21.7588 402.406 21.9053 402.783 22.1982C403.164 22.488 403.446 22.9079 403.628 23.458C403.81 24.0049 403.901 24.6657 403.901 25.4404ZM400.664 25.4404C400.664 25.9059 400.68 26.2965 400.713 26.6123C400.749 26.9281 400.814 27.1657 400.908 27.3252C401.006 27.4814 401.144 27.5596 401.323 27.5596C401.502 27.5596 401.639 27.4814 401.733 27.3252C401.828 27.1657 401.893 26.9297 401.929 26.6172C401.965 26.3014 401.982 25.9092 401.982 25.4404C401.982 24.9684 401.965 24.5745 401.929 24.2588C401.893 23.943 401.828 23.7054 401.733 23.5459C401.639 23.3831 401.502 23.3018 401.323 23.3018C401.144 23.3018 401.006 23.3831 400.908 23.5459C400.814 23.7054 400.749 23.943 400.713 24.2588C400.68 24.5745 400.664 24.9684 400.664 25.4404Z" fill="#F5F5F5" />
-          </g>
-        </g>
-      </g>
-      <defs>
-        <filter id="filter0_f_311_518" x="247.8" y="13.9777" width="1.82514" height="3.5538" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="0.101426" result="effect1_foregroundBlur_311_518" />
-        </filter>
-        <filter id="filter1_f_311_518" x="250.617" y="13.5392" width="2.91449" height="3.55954" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="0.101426" result="effect1_foregroundBlur_311_518" />
-        </filter>
-        <clipPath id="clip0_311_518">
-          <rect width="428" height="55" fill="white" />
-        </clipPath>
-        <clipPath id="clip1_311_518">
-          <rect width="211" height="33" fill="white" transform="translate(108.5)" />
-        </clipPath>
-        <clipPath id="clip2_311_518">
-          <rect width="211" height="33" fill="white" transform="translate(108)" />
-        </clipPath>
-        <clipPath id="clip3_311_518">
-          <rect width="83.4286" height="34" fill="white" transform="translate(328.036 8)" />
-        </clipPath>
-        <clipPath id="clip4_311_518">
-          <rect width="22.8571" height="16" fill="white" transform="translate(328.036 17)" />
-        </clipPath>
-        <clipPath id="clip5_311_518">
-          <rect width="29.7143" height="16" fill="white" transform="translate(381.75 17)" />
-        </clipPath>
-      </defs>
-    </svg>
-
-  )
-}
-
-export default StatusBar;
