+++ src/frames/iphone-15-pro/svg/status-bar.tsx
@@ -1,0 +1,55 @@
+interface StatusBarProps extends App.SVGProps {
+  foreground?: /*dark*/'#030712' | /*light*/'#f3f4f6',
+  bPercentage?: /*dark*/'#e2e8f0' | /*light*/'transparent'
+}
+
+
+const StatusBar: Nixix.FC<StatusBarProps> = ({ className, foreground = '#030712', bPercentage = '#e2e8f0', ...rest }): someView => {
+  return (
+    <svg
+      {...rest} className={className + ' tws-h-fit'}
+      width="393" height="54" viewBox="0 0 393 54" fill="none" xmlns="http://www.w3.org/2000/svg">
+      <g clip-path="url(#clip0_317_1609)">
+        <path d="M61.7666 25.7231C62.4085 25.7231 63.0173 25.8421 63.5928 26.0801C64.1683 26.3125 64.6774 26.6777 65.1201 27.1758C65.5684 27.6683 65.9198 28.3019 66.1743 29.0767C66.4289 29.8459 66.5562 30.7728 66.5562 31.8574V31.874C66.5562 33.2188 66.3625 34.3698 65.9751 35.3271C65.5877 36.2845 65.0316 37.0205 64.3066 37.5352C63.5872 38.0443 62.7267 38.2988 61.7251 38.2988C60.9891 38.2988 60.325 38.166 59.7329 37.9004C59.1463 37.6292 58.6593 37.2585 58.272 36.7881C57.8901 36.3122 57.6328 35.7699 57.5 35.1611L57.4834 35.0698H59.5835L59.6167 35.1528C59.7218 35.4295 59.8713 35.673 60.0649 35.8833C60.2642 36.0936 60.5021 36.2568 60.7788 36.373C61.061 36.4837 61.3765 36.5391 61.7251 36.5391C62.356 36.5391 62.8706 36.3537 63.269 35.9829C63.673 35.6066 63.9746 35.1058 64.1738 34.4805C64.3786 33.8496 64.4948 33.1523 64.5225 32.3887C64.528 32.3057 64.5308 32.2254 64.5308 32.1479C64.5308 32.0649 64.5308 31.9847 64.5308 31.9072L64.1572 29.9067C64.1572 29.453 64.0521 29.0407 63.8418 28.6699C63.6315 28.2992 63.3465 28.0059 62.9868 27.79C62.6326 27.5742 62.2314 27.4663 61.7832 27.4663C61.3516 27.4663 60.9559 27.5715 60.5962 27.7817C60.2365 27.992 59.9487 28.2798 59.7329 28.645C59.5171 29.0047 59.4092 29.4087 59.4092 29.8569V29.8735C59.4092 30.3384 59.5116 30.7507 59.7163 31.1104C59.9211 31.4645 60.2005 31.744 60.5547 31.9487C60.9089 32.1535 61.3128 32.2559 61.7666 32.2559C62.2204 32.2559 62.6271 32.1562 62.9868 31.957C63.3465 31.7578 63.6315 31.4811 63.8418 31.127C64.0521 30.7728 64.1572 30.3716 64.1572 29.9233V29.9067H64.6553V32.1147H64.3647C64.2209 32.4302 64.0078 32.7235 63.7256 32.9946C63.4434 33.2658 63.0947 33.4844 62.6797 33.6504C62.2702 33.8164 61.7887 33.8994 61.2354 33.8994C60.4717 33.8994 59.7965 33.7279 59.21 33.3848C58.6234 33.0417 58.1613 32.5713 57.8237 31.9736C57.4917 31.376 57.3257 30.7008 57.3257 29.9482V29.9316C57.3257 29.1182 57.5138 28.396 57.8901 27.7651C58.272 27.1287 58.7977 26.6307 59.4673 26.271C60.1424 25.9058 60.9089 25.7231 61.7666 25.7231ZM70.3247 36.3896C69.9484 36.3896 69.633 36.2651 69.3784 36.0161C69.1294 35.7616 69.0049 35.4517 69.0049 35.0864C69.0049 34.7157 69.1294 34.4058 69.3784 34.1567C69.633 33.9022 69.9484 33.7749 70.3247 33.7749C70.7065 33.7749 71.022 33.9022 71.271 34.1567C71.52 34.4058 71.6445 34.7157 71.6445 35.0864C71.6445 35.4517 71.52 35.7616 71.271 36.0161C71.022 36.2651 70.7065 36.3896 70.3247 36.3896ZM70.3247 30.2305C69.9484 30.2305 69.633 30.106 69.3784 29.8569C69.1294 29.6024 69.0049 29.2925 69.0049 28.9272C69.0049 28.5565 69.1294 28.2466 69.3784 27.9976C69.633 27.743 69.9484 27.6157 70.3247 27.6157C70.7065 27.6157 71.022 27.743 71.271 27.9976C71.52 28.2466 71.6445 28.5565 71.6445 28.9272C71.6445 29.2925 71.52 29.6024 71.271 29.8569C71.022 30.106 70.7065 30.2305 70.3247 30.2305ZM79.9038 38V35.7007H74.0435V33.9492C74.3589 33.4124 74.6743 32.8729 74.9897 32.3306C75.3107 31.7882 75.6344 31.2487 75.9609 30.7119C76.293 30.1696 76.6222 29.6356 76.9487 29.1099C77.2808 28.5786 77.61 28.0557 77.9365 27.541C78.2686 27.0208 78.6006 26.5145 78.9326 26.022H81.9458V33.9326H83.5645V35.7007H81.9458V38H79.9038ZM76.0356 33.9824H79.937V27.7319H79.8125C79.569 28.1082 79.3172 28.4956 79.0571 28.894C78.8026 29.2925 78.5452 29.6992 78.2852 30.1143C78.0251 30.5293 77.7677 30.9471 77.5132 31.3677C77.2586 31.7827 77.0068 32.2005 76.7578 32.6211C76.5088 33.0361 76.2681 33.4484 76.0356 33.8579V33.9824ZM88.52 38V28.1304H88.3706L85.3906 30.2305V28.2217L88.52 26.022H90.6533V38H88.52Z" fill="black" />
+        <path fill-rule="evenodd" clip-rule="evenodd" d="M307.865 27.033C307.865 26.4 307.387 25.8868 306.798 25.8868H305.732C305.143 25.8868 304.665 26.4 304.665 27.033V36.967C304.665 37.6 305.143 38.1132 305.732 38.1132H306.798C307.387 38.1132 307.865 37.6 307.865 36.967V27.033ZM300.431 28.3321H301.498C302.087 28.3321 302.564 28.8576 302.564 29.5058V36.9395C302.564 37.5877 302.087 38.1132 301.498 38.1132H300.431C299.842 38.1132 299.364 37.5877 299.364 36.9395V29.5058C299.364 28.8576 299.842 28.3321 300.431 28.3321ZM296.099 30.9811H295.032C294.443 30.9811 293.966 31.5133 293.966 32.1698V36.9245C293.966 37.581 294.443 38.1132 295.032 38.1132H296.099C296.688 38.1132 297.166 37.581 297.166 36.9245V32.1698C297.166 31.5133 296.688 30.9811 296.099 30.9811ZM290.798 33.4264H289.732C289.143 33.4264 288.665 33.951 288.665 34.5981V36.9415C288.665 37.5886 289.143 38.1132 289.732 38.1132H290.798C291.387 38.1132 291.865 37.5886 291.865 36.9415V34.5981C291.865 33.951 291.387 33.4264 290.798 33.4264Z" fill="black" />
+        <path fill-rule="evenodd" clip-rule="evenodd" d="M323.437 28.3021C325.924 28.3022 328.316 29.2243 330.118 30.8778C330.254 31.0054 330.471 31.0038 330.605 30.8742L331.902 29.6107C331.97 29.545 332.007 29.4559 332.007 29.3632C332.006 29.2705 331.968 29.1819 331.899 29.1169C327.168 24.7422 319.704 24.7422 314.973 29.1169C314.905 29.1818 314.866 29.2704 314.865 29.3631C314.865 29.4558 314.902 29.5449 314.97 29.6107L316.268 30.8742C316.401 31.004 316.619 31.0056 316.754 30.8778C318.557 29.2242 320.949 28.3021 323.437 28.3021ZM323.433 32.5224C324.791 32.5223 326.099 33.0341 327.106 33.9582C327.242 34.0893 327.456 34.0865 327.589 33.9518L328.876 32.6325C328.944 32.5633 328.981 32.4694 328.98 32.3718C328.979 32.2743 328.94 32.1812 328.871 32.1134C325.807 29.2226 321.062 29.2226 317.998 32.1134C317.929 32.1812 317.89 32.2743 317.889 32.3719C317.888 32.4695 317.926 32.5634 317.993 32.6325L319.28 33.9518C319.413 34.0865 319.627 34.0893 319.763 33.9582C320.769 33.0347 322.077 32.523 323.433 32.5224ZM325.958 35.316C325.96 35.4213 325.922 35.5229 325.855 35.5967L323.679 38.0514C323.615 38.1236 323.528 38.1642 323.437 38.1642C323.346 38.1642 323.259 38.1236 323.195 38.0514L321.018 35.5967C320.951 35.5228 320.914 35.4212 320.916 35.3159C320.918 35.2105 320.959 35.1108 321.029 35.0401C322.419 33.7262 324.455 33.7262 325.845 35.0401C325.915 35.1108 325.956 35.2106 325.958 35.316Z" fill="black" />
+        <rect opacity="0.35" x="339.507" y="26" width="24" height="12" rx="3.8" stroke="black" />
+        <path opacity="0.4" d="M365.007 30.2811V34.3566C365.812 34.0114 366.335 33.2085 366.335 32.3189C366.335 31.4293 365.812 30.6263 365.007 30.2811Z" fill="black" />
+        <rect x="341.007" y="27.5" width="21" height="9" rx="2.5" fill="black" />
+        <g clip-path="url(#clip1_317_1609)">
+          <rect x="139" y="15" width="127" height="35.3113" rx="17.6556" fill="black" />
+          <path d="M247.347 39.1909C251.36 39.1909 254.613 36.0964 254.613 32.2792C254.613 28.4619 251.36 25.3675 247.347 25.3675C243.334 25.3675 240.081 28.4619 240.081 32.2792C240.081 36.0964 243.334 39.1909 247.347 39.1909Z" fill="#161616" />
+          <path d="M247.36 36.5417C249.835 36.5417 251.841 34.633 251.841 32.2784C251.841 29.9239 249.835 28.0152 247.36 28.0152C244.884 28.0152 242.878 29.9239 242.878 32.2784C242.878 34.633 244.884 36.5417 247.36 36.5417Z" fill="#0A0D13" />
+          <path d="M247.35 35.5268C249.236 35.5268 250.765 34.0725 250.765 32.2786C250.765 30.4846 249.236 29.0303 247.35 29.0303C245.464 29.0303 243.936 30.4846 243.936 32.2786C243.936 34.0725 245.464 35.5268 247.35 35.5268Z" fill="#091427" />
+          <g filter="url(#filter0_f_317_1609)">
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M245.006 30.1055C244.601 30.1461 244.026 30.734 243.983 31.8895C243.94 33.0451 244.473 33.6736 244.771 33.6736C245.07 33.6736 246.306 31.9909 245.006 30.1055Z" fill="#235A91" fill-opacity="0.556075" />
+          </g>
+          <g filter="url(#filter1_f_317_1609)">
+            <path fill-rule="evenodd" clip-rule="evenodd" d="M247.907 29.7011C247.268 29.8881 247.145 30.8696 247.17 31.4304C247.194 31.9913 248.03 33.4635 249.037 33.1363C250.044 32.8092 250.216 31.594 249.798 30.7761C249.381 29.9582 248.619 29.374 247.907 29.7011Z" fill="#235A91" fill-opacity="0.556075" />
+          </g>
+        </g>
+      </g>
+      <defs>
+        <filter id="filter0_f_317_1609" x="243.805" y="29.9303" width="1.9589" height="3.91853" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
+          <feFlood flood-opacity="0" result="BackgroundImageFix" />
+          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
+          <feGaussianBlur stdDeviation="0.0876262" result="effect1_foregroundBlur_317_1609" />
+        </filter>
+        <filter id="filter1_f_317_1609" x="246.992" y="29.4334" width="3.19425" height="3.92506" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
+          <feFlood flood-opacity="0" result="BackgroundImageFix" />
+          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
+          <feGaussianBlur stdDeviation="0.0876262" result="effect1_foregroundBlur_317_1609" />
+        </filter>
+        <clipPath id="clip0_317_1609">
+          <rect width="393" height="54" fill="white" />
+        </clipPath>
+        <clipPath id="clip1_317_1609">
+          <rect width="127" height="35.3095" fill="white" transform="translate(139 15)" />
+        </clipPath>
+      </defs>
+    </svg>
+  )
+};
+
+
+export default StatusBar;
