+++ src/frames/iphone-16-pro/svg/device-image.png
