+++ src/frames/iphone-16-pro/svg/status-bar.tsx
