--- src/frames/iphone-5s/svg/status-bar.tsx
@@ -1,26 +1,0 @@
-const StatusBar: Nixix.FC<App.SVGProps> = ({ className, ...rest }): someView => {
-  return (
-    <svg
-      {...rest} className={className + ' tws-h-fit '}
-      width="375" height="20" viewBox="0 0 375 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <g clip-path="url(#clip0_288_439)">
-        <path fill-rule="evenodd" clip-rule="evenodd" d="M21.6492 4C21.097 4 20.6492 4.44772 20.6492 5V15C20.6492 15.5523 21.0969 16 21.6492 16H23C23.5523 16 24 15.5523 24 15V5C24 4.44772 23.5523 4 23 4H21.6492ZM16.0995 6.40002C15.5472 6.40002 15.0995 6.84774 15.0995 7.40002V15C15.0995 15.5523 15.5472 16 16.0995 16H17.4503C18.0026 16 18.4503 15.5523 18.4503 15V7.40002C18.4503 6.84774 18.0026 6.40002 17.4503 6.40002H16.0995ZM9.54983 10C9.54983 9.44772 9.99755 9 10.5498 9H11.9006C12.4529 9 12.9006 9.44772 12.9006 10V15C12.9006 15.5523 12.4529 16 11.9006 16H10.5498C9.99754 16 9.54983 15.5523 9.54983 15V10ZM5 11.4C4.44772 11.4 4 11.8477 4 12.4V15C4 15.5523 4.44771 16 5 16H6.35078C6.90307 16 7.35078 15.5523 7.35078 15V12.4C7.35078 11.8477 6.90307 11.4 6.35078 11.4H5Z" fill="#111111" />
-      </g>
-      <path d="M36.2285 11.2871H37.3535C37.2949 11.8262 37.1406 12.3086 36.8906 12.7344C36.6406 13.1602 36.2871 13.498 35.8301 13.748C35.373 13.9941 34.8027 14.1172 34.1191 14.1172C33.6191 14.1172 33.1641 14.0234 32.7539 13.8359C32.3477 13.6484 31.998 13.3828 31.7051 13.0391C31.4121 12.6914 31.1855 12.2754 31.0254 11.791C30.8691 11.3027 30.791 10.7598 30.791 10.1621V9.3125C30.791 8.71484 30.8691 8.17383 31.0254 7.68945C31.1855 7.20117 31.4141 6.7832 31.7109 6.43555C32.0117 6.08789 32.373 5.82031 32.7949 5.63281C33.2168 5.44531 33.6914 5.35156 34.2188 5.35156C34.8633 5.35156 35.4082 5.47266 35.8535 5.71484C36.2988 5.95703 36.6445 6.29297 36.8906 6.72266C37.1406 7.14844 37.2949 7.64258 37.3535 8.20508H36.2285C36.1738 7.80664 36.0723 7.46484 35.9238 7.17969C35.7754 6.89062 35.5645 6.66797 35.291 6.51172C35.0176 6.35547 34.6602 6.27734 34.2188 6.27734C33.8398 6.27734 33.5059 6.34961 33.2168 6.49414C32.9316 6.63867 32.6914 6.84375 32.4961 7.10938C32.3047 7.375 32.1602 7.69336 32.0625 8.06445C31.9648 8.43555 31.916 8.84766 31.916 9.30078V10.1621C31.916 10.5801 31.959 10.9727 32.0449 11.3398C32.1348 11.707 32.2695 12.0293 32.4492 12.3066C32.6289 12.584 32.8574 12.8027 33.1348 12.9629C33.4121 13.1191 33.7402 13.1973 34.1191 13.1973C34.5996 13.1973 34.9824 13.1211 35.2676 12.9688C35.5527 12.8164 35.7676 12.5977 35.9121 12.3125C36.0605 12.0273 36.166 11.6855 36.2285 11.2871ZM42.4688 12.916V9.65234C42.4688 9.40234 42.418 9.18555 42.3164 9.00195C42.2188 8.81445 42.0703 8.66992 41.8711 8.56836C41.6719 8.4668 41.4258 8.41602 41.1328 8.41602C40.8594 8.41602 40.6191 8.46289 40.4121 8.55664C40.209 8.65039 40.0488 8.77344 39.9316 8.92578C39.8184 9.07812 39.7617 9.24219 39.7617 9.41797H38.6777C38.6777 9.19141 38.7363 8.9668 38.8535 8.74414C38.9707 8.52148 39.1387 8.32031 39.3574 8.14062C39.5801 7.95703 39.8457 7.8125 40.1543 7.70703C40.4668 7.59766 40.8145 7.54297 41.1973 7.54297C41.6582 7.54297 42.0645 7.62109 42.416 7.77734C42.7715 7.93359 43.0488 8.16992 43.248 8.48633C43.4512 8.79883 43.5527 9.19141 43.5527 9.66406V12.6172C43.5527 12.8281 43.5703 13.0527 43.6055 13.291C43.6445 13.5293 43.7012 13.7344 43.7754 13.9062V14H42.6445C42.5898 13.875 42.5469 13.709 42.5156 13.502C42.4844 13.291 42.4688 13.0957 42.4688 12.916ZM42.6562 10.1562L42.668 10.918H41.5723C41.2637 10.918 40.9883 10.9434 40.7461 10.9941C40.5039 11.041 40.3008 11.1133 40.1367 11.2109C39.9727 11.3086 39.8477 11.4316 39.7617 11.5801C39.6758 11.7246 39.6328 11.8945 39.6328 12.0898C39.6328 12.2891 39.6777 12.4707 39.7676 12.6348C39.8574 12.7988 39.9922 12.9297 40.1719 13.0273C40.3555 13.1211 40.5801 13.168 40.8457 13.168C41.1777 13.168 41.4707 13.0977 41.7246 12.957C41.9785 12.8164 42.1797 12.6445 42.3281 12.4414C42.4805 12.2383 42.5625 12.041 42.5742 11.8496L43.0371 12.3711C43.0098 12.5352 42.9355 12.7168 42.8145 12.916C42.6934 13.1152 42.5312 13.3066 42.3281 13.4902C42.1289 13.6699 41.8906 13.8203 41.6133 13.9414C41.3398 14.0586 41.0312 14.1172 40.6875 14.1172C40.2578 14.1172 39.8809 14.0332 39.5566 13.8652C39.2363 13.6973 38.9863 13.4727 38.8066 13.1914C38.6309 12.9062 38.543 12.5879 38.543 12.2363C38.543 11.8965 38.6094 11.5977 38.7422 11.3398C38.875 11.0781 39.0664 10.8613 39.3164 10.6895C39.5664 10.5137 39.8672 10.3809 40.2188 10.291C40.5703 10.2012 40.9629 10.1562 41.3965 10.1562H42.6562ZM46.3418 8.65625V14H45.2578V7.66016H46.3125L46.3418 8.65625ZM48.3223 7.625L48.3164 8.63281C48.2266 8.61328 48.1406 8.60156 48.0586 8.59766C47.9805 8.58984 47.8906 8.58594 47.7891 8.58594C47.5391 8.58594 47.3184 8.625 47.127 8.70312C46.9355 8.78125 46.7734 8.89062 46.6406 9.03125C46.5078 9.17188 46.4023 9.33984 46.3242 9.53516C46.25 9.72656 46.2012 9.9375 46.1777 10.168L45.873 10.3438C45.873 9.96094 45.9102 9.60156 45.9844 9.26562C46.0625 8.92969 46.1816 8.63281 46.3418 8.375C46.502 8.11328 46.7051 7.91016 46.9512 7.76562C47.2012 7.61719 47.498 7.54297 47.8418 7.54297C47.9199 7.54297 48.0098 7.55273 48.1113 7.57227C48.2129 7.58789 48.2832 7.60547 48.3223 7.625ZM50.4082 8.65625V14H49.3242V7.66016H50.3789L50.4082 8.65625ZM52.3887 7.625L52.3828 8.63281C52.293 8.61328 52.207 8.60156 52.125 8.59766C52.0469 8.58984 51.957 8.58594 51.8555 8.58594C51.6055 8.58594 51.3848 8.625 51.1934 8.70312C51.002 8.78125 50.8398 8.89062 50.707 9.03125C50.5742 9.17188 50.4688 9.33984 50.3906 9.53516C50.3164 9.72656 50.2676 9.9375 50.2441 10.168L49.9395 10.3438C49.9395 9.96094 49.9766 9.60156 50.0508 9.26562C50.1289 8.92969 50.248 8.63281 50.4082 8.375C50.5684 8.11328 50.7715 7.91016 51.0176 7.76562C51.2676 7.61719 51.5645 7.54297 51.9082 7.54297C51.9863 7.54297 52.0762 7.55273 52.1777 7.57227C52.2793 7.58789 52.3496 7.60547 52.3887 7.625ZM54.5684 7.66016V14H53.4785V7.66016H54.5684ZM53.3965 5.97852C53.3965 5.80273 53.4492 5.6543 53.5547 5.5332C53.6641 5.41211 53.8242 5.35156 54.0352 5.35156C54.2422 5.35156 54.4004 5.41211 54.5098 5.5332C54.623 5.6543 54.6797 5.80273 54.6797 5.97852C54.6797 6.14648 54.623 6.29102 54.5098 6.41211C54.4004 6.5293 54.2422 6.58789 54.0352 6.58789C53.8242 6.58789 53.6641 6.5293 53.5547 6.41211C53.4492 6.29102 53.3965 6.14648 53.3965 5.97852ZM58.9395 14.1172C58.498 14.1172 58.0977 14.043 57.7383 13.8945C57.3828 13.7422 57.0762 13.5293 56.8184 13.2559C56.5645 12.9824 56.3691 12.6582 56.2324 12.2832C56.0957 11.9082 56.0273 11.498 56.0273 11.0527V10.8066C56.0273 10.291 56.1035 9.83203 56.2559 9.42969C56.4082 9.02344 56.6152 8.67969 56.877 8.39844C57.1387 8.11719 57.4355 7.9043 57.7676 7.75977C58.0996 7.61523 58.4434 7.54297 58.7988 7.54297C59.252 7.54297 59.6426 7.62109 59.9707 7.77734C60.3027 7.93359 60.5742 8.15234 60.7852 8.43359C60.9961 8.71094 61.1523 9.03906 61.2539 9.41797C61.3555 9.79297 61.4062 10.2031 61.4062 10.6484V11.1348H56.6719V10.25H60.3223V10.168C60.3066 9.88672 60.248 9.61328 60.1465 9.34766C60.0488 9.08203 59.8926 8.86328 59.6777 8.69141C59.4629 8.51953 59.1699 8.43359 58.7988 8.43359C58.5527 8.43359 58.3262 8.48633 58.1191 8.5918C57.9121 8.69336 57.7344 8.8457 57.5859 9.04883C57.4375 9.25195 57.3223 9.5 57.2402 9.79297C57.1582 10.0859 57.1172 10.4238 57.1172 10.8066V11.0527C57.1172 11.3535 57.1582 11.6367 57.2402 11.9023C57.3262 12.1641 57.4492 12.3945 57.6094 12.5938C57.7734 12.793 57.9707 12.9492 58.2012 13.0625C58.4355 13.1758 58.7012 13.2324 58.998 13.2324C59.3809 13.2324 59.7051 13.1543 59.9707 12.998C60.2363 12.8418 60.4688 12.6328 60.668 12.3711L61.3242 12.8926C61.1875 13.0996 61.0137 13.2969 60.8027 13.4844C60.5918 13.6719 60.332 13.8242 60.0234 13.9414C59.7188 14.0586 59.3574 14.1172 58.9395 14.1172ZM63.7559 8.65625V14H62.6719V7.66016H63.7266L63.7559 8.65625ZM65.7363 7.625L65.7305 8.63281C65.6406 8.61328 65.5547 8.60156 65.4727 8.59766C65.3945 8.58984 65.3047 8.58594 65.2031 8.58594C64.9531 8.58594 64.7324 8.625 64.541 8.70312C64.3496 8.78125 64.1875 8.89062 64.0547 9.03125C63.9219 9.17188 63.8164 9.33984 63.7383 9.53516C63.6641 9.72656 63.6152 9.9375 63.5918 10.168L63.2871 10.3438C63.2871 9.96094 63.3242 9.60156 63.3984 9.26562C63.4766 8.92969 63.5957 8.63281 63.7559 8.375C63.916 8.11328 64.1191 7.91016 64.3652 7.76562C64.6152 7.61719 64.9121 7.54297 65.2559 7.54297C65.334 7.54297 65.4238 7.55273 65.5254 7.57227C65.627 7.58789 65.6973 7.60547 65.7363 7.625Z" fill="#111111" />
-      <path d="M82.0089 12.3275C82.7269 12.3275 83.4246 12.5292 84.0394 12.9137L84.2885 13.0695C84.4773 13.1876 84.5105 13.464 84.3557 13.629L82.229 15.8956C82.0984 16.0348 81.8866 16.0348 81.7559 15.8956L79.6429 13.6436C79.4888 13.4793 79.5209 13.2043 79.7081 13.0853L79.9544 12.9288C80.5748 12.5345 81.2814 12.3275 82.0089 12.3275Z" fill="#111111" />
-      <path d="M82.0086 8.16389C83.7948 8.16389 85.5098 8.79718 86.9069 9.96997L87.1044 10.1357C87.2618 10.2679 87.2756 10.5168 87.1339 10.6679L85.8643 12.021C85.7469 12.1462 85.561 12.1606 85.4279 12.0548L85.2731 11.9318C84.3295 11.1824 83.1917 10.7803 82.0086 10.7803C80.8183 10.7803 79.6738 11.1874 78.7271 11.9454L78.5722 12.0695C78.439 12.1761 78.2525 12.1621 78.1347 12.0366L76.8656 10.6839C76.7242 10.5332 76.7376 10.2849 76.8943 10.1525L77.0906 9.98659C78.491 8.80334 80.2138 8.16389 82.0086 8.16389Z" fill="#111111" />
-      <path d="M82.0087 4C84.8434 4 87.5542 5.06869 89.7 7.02754L89.8831 7.19474C90.0311 7.32986 90.0398 7.5707 89.902 7.71757L88.6365 9.06634C88.5139 9.19702 88.3179 9.20617 88.1851 9.08744L88.0284 8.9474C86.339 7.43766 84.2215 6.61643 82.0087 6.61643C79.7883 6.61643 77.664 7.44332 75.9721 8.9625L75.8154 9.10325C75.6825 9.22252 75.4861 9.21359 75.3633 9.0827L74.098 7.73412C73.9604 7.58745 73.9688 7.34702 74.1164 7.21176L74.2988 7.0445C76.4472 5.07507 79.1655 4 82.0087 4Z" fill="#111111" />
-      <path d="M166.002 5.44531V14H164.449V7.25586L162.393 7.93555V6.68164L165.826 5.44531H166.002ZM174.521 9.00195V10.4375C174.521 11.0938 174.455 11.6543 174.322 12.1191C174.189 12.584 173.998 12.9648 173.748 13.2617C173.498 13.5547 173.199 13.7715 172.852 13.9121C172.504 14.0488 172.117 14.1172 171.691 14.1172C171.352 14.1172 171.035 14.0742 170.742 13.9883C170.449 13.8984 170.186 13.7617 169.951 13.5781C169.721 13.3906 169.521 13.1523 169.354 12.8633C169.189 12.5742 169.062 12.2285 168.973 11.8262C168.887 11.4238 168.844 10.9609 168.844 10.4375V9.00195C168.844 8.3457 168.91 7.78711 169.043 7.32617C169.18 6.86523 169.373 6.48828 169.623 6.19531C169.873 5.90234 170.172 5.68945 170.52 5.55664C170.867 5.41992 171.254 5.35156 171.68 5.35156C172.023 5.35156 172.34 5.39453 172.629 5.48047C172.922 5.56641 173.184 5.70117 173.414 5.88477C173.648 6.06445 173.848 6.29883 174.012 6.58789C174.176 6.87305 174.301 7.2168 174.387 7.61914C174.477 8.01758 174.521 8.47852 174.521 9.00195ZM172.975 10.6484V8.78516C172.975 8.45703 172.955 8.17188 172.916 7.92969C172.881 7.68359 172.828 7.47461 172.758 7.30273C172.688 7.13086 172.6 6.99219 172.494 6.88672C172.389 6.77734 172.268 6.69727 172.131 6.64648C171.994 6.5957 171.844 6.57031 171.68 6.57031C171.477 6.57031 171.295 6.61133 171.135 6.69336C170.975 6.77148 170.84 6.89648 170.73 7.06836C170.621 7.24023 170.537 7.46875 170.479 7.75391C170.424 8.03516 170.396 8.37891 170.396 8.78516V10.6484C170.396 10.9766 170.414 11.2637 170.449 11.5098C170.488 11.7559 170.543 11.9668 170.613 12.1426C170.688 12.3184 170.775 12.4629 170.877 12.5762C170.982 12.6855 171.104 12.7676 171.24 12.8223C171.377 12.873 171.527 12.8984 171.691 12.8984C171.898 12.8984 172.08 12.8574 172.236 12.7754C172.396 12.6934 172.531 12.5645 172.641 12.3887C172.75 12.209 172.832 11.9766 172.887 11.6914C172.945 11.4023 172.975 11.0547 172.975 10.6484ZM175.846 13.2441C175.846 13.0098 175.926 12.8125 176.086 12.6523C176.246 12.4922 176.463 12.4121 176.736 12.4121C177.01 12.4121 177.227 12.4922 177.387 12.6523C177.547 12.8125 177.627 13.0098 177.627 13.2441C177.627 13.4746 177.547 13.6699 177.387 13.8301C177.227 13.9863 177.01 14.0645 176.736 14.0645C176.463 14.0645 176.246 13.9863 176.086 13.8301C175.926 13.6699 175.846 13.4746 175.846 13.2441ZM175.852 8.32812C175.852 8.09375 175.932 7.89648 176.092 7.73633C176.252 7.57617 176.469 7.49609 176.742 7.49609C177.016 7.49609 177.232 7.57617 177.393 7.73633C177.553 7.89648 177.633 8.09375 177.633 8.32812C177.633 8.55859 177.553 8.75391 177.393 8.91406C177.232 9.07031 177.016 9.14844 176.742 9.14844C176.469 9.14844 176.252 9.07031 176.092 8.91406C175.932 8.75391 175.852 8.55859 175.852 8.32812ZM184.646 9.00195V10.4375C184.646 11.0938 184.58 11.6543 184.447 12.1191C184.314 12.584 184.123 12.9648 183.873 13.2617C183.623 13.5547 183.324 13.7715 182.977 13.9121C182.629 14.0488 182.242 14.1172 181.816 14.1172C181.477 14.1172 181.16 14.0742 180.867 13.9883C180.574 13.8984 180.311 13.7617 180.076 13.5781C179.846 13.3906 179.646 13.1523 179.479 12.8633C179.314 12.5742 179.188 12.2285 179.098 11.8262C179.012 11.4238 178.969 10.9609 178.969 10.4375V9.00195C178.969 8.3457 179.035 7.78711 179.168 7.32617C179.305 6.86523 179.498 6.48828 179.748 6.19531C179.998 5.90234 180.297 5.68945 180.645 5.55664C180.992 5.41992 181.379 5.35156 181.805 5.35156C182.148 5.35156 182.465 5.39453 182.754 5.48047C183.047 5.56641 183.309 5.70117 183.539 5.88477C183.773 6.06445 183.973 6.29883 184.137 6.58789C184.301 6.87305 184.426 7.2168 184.512 7.61914C184.602 8.01758 184.646 8.47852 184.646 9.00195ZM183.1 10.6484V8.78516C183.1 8.45703 183.08 8.17188 183.041 7.92969C183.006 7.68359 182.953 7.47461 182.883 7.30273C182.812 7.13086 182.725 6.99219 182.619 6.88672C182.514 6.77734 182.393 6.69727 182.256 6.64648C182.119 6.5957 181.969 6.57031 181.805 6.57031C181.602 6.57031 181.42 6.61133 181.26 6.69336C181.1 6.77148 180.965 6.89648 180.855 7.06836C180.746 7.24023 180.662 7.46875 180.604 7.75391C180.549 8.03516 180.521 8.37891 180.521 8.78516V10.6484C180.521 10.9766 180.539 11.2637 180.574 11.5098C180.613 11.7559 180.668 11.9668 180.738 12.1426C180.812 12.3184 180.9 12.4629 181.002 12.5762C181.107 12.6855 181.229 12.7676 181.365 12.8223C181.502 12.873 181.652 12.8984 181.816 12.8984C182.023 12.8984 182.205 12.8574 182.361 12.7754C182.521 12.6934 182.656 12.5645 182.766 12.3887C182.875 12.209 182.957 11.9766 183.012 11.6914C183.07 11.4023 183.1 11.0547 183.1 10.6484ZM191.502 9.00195V10.4375C191.502 11.0938 191.436 11.6543 191.303 12.1191C191.17 12.584 190.979 12.9648 190.729 13.2617C190.479 13.5547 190.18 13.7715 189.832 13.9121C189.484 14.0488 189.098 14.1172 188.672 14.1172C188.332 14.1172 188.016 14.0742 187.723 13.9883C187.43 13.8984 187.166 13.7617 186.932 13.5781C186.701 13.3906 186.502 13.1523 186.334 12.8633C186.17 12.5742 186.043 12.2285 185.953 11.8262C185.867 11.4238 185.824 10.9609 185.824 10.4375V9.00195C185.824 8.3457 185.891 7.78711 186.023 7.32617C186.16 6.86523 186.354 6.48828 186.604 6.19531C186.854 5.90234 187.152 5.68945 187.5 5.55664C187.848 5.41992 188.234 5.35156 188.66 5.35156C189.004 5.35156 189.32 5.39453 189.609 5.48047C189.902 5.56641 190.164 5.70117 190.395 5.88477C190.629 6.06445 190.828 6.29883 190.992 6.58789C191.156 6.87305 191.281 7.2168 191.367 7.61914C191.457 8.01758 191.502 8.47852 191.502 9.00195ZM189.955 10.6484V8.78516C189.955 8.45703 189.936 8.17188 189.896 7.92969C189.861 7.68359 189.809 7.47461 189.738 7.30273C189.668 7.13086 189.58 6.99219 189.475 6.88672C189.369 6.77734 189.248 6.69727 189.111 6.64648C188.975 6.5957 188.824 6.57031 188.66 6.57031C188.457 6.57031 188.275 6.61133 188.115 6.69336C187.955 6.77148 187.82 6.89648 187.711 7.06836C187.602 7.24023 187.518 7.46875 187.459 7.75391C187.404 8.03516 187.377 8.37891 187.377 8.78516V10.6484C187.377 10.9766 187.395 11.2637 187.43 11.5098C187.469 11.7559 187.523 11.9668 187.594 12.1426C187.668 12.3184 187.756 12.4629 187.857 12.5762C187.963 12.6855 188.084 12.7676 188.221 12.8223C188.357 12.873 188.508 12.8984 188.672 12.8984C188.879 12.8984 189.061 12.8574 189.217 12.7754C189.377 12.6934 189.512 12.5645 189.621 12.3887C189.73 12.209 189.812 11.9766 189.867 11.6914C189.926 11.4023 189.955 11.0547 189.955 10.6484ZM199.289 6.76367L196.857 14H195.152L198.34 5.46875H199.43L199.289 6.76367ZM201.322 14L198.879 6.76367L198.732 5.46875H199.828L203.033 14H201.322ZM201.211 10.8301V12.1016H196.635V10.8301H201.211ZM204.615 5.46875H206.01L208.342 11.8906L210.668 5.46875H212.068L208.898 14H207.779L204.615 5.46875ZM203.912 5.46875H205.277L205.523 11.3691V14H203.912V5.46875ZM211.4 5.46875H212.771V14H211.16V11.3691L211.4 5.46875Z" fill="#111111" />
-      <path d="M311.752 5.42188V14H310.668V6.77539L308.482 7.57227V6.59375L311.582 5.42188H311.752ZM320.295 9.05469V10.3555C320.295 11.0547 320.232 11.6445 320.107 12.125C319.982 12.6055 319.803 12.9922 319.568 13.2852C319.334 13.5781 319.051 13.791 318.719 13.9238C318.391 14.0527 318.02 14.1172 317.605 14.1172C317.277 14.1172 316.975 14.0762 316.697 13.9941C316.42 13.9121 316.17 13.7812 315.947 13.6016C315.729 13.418 315.541 13.1797 315.385 12.8867C315.229 12.5938 315.109 12.2383 315.027 11.8203C314.945 11.4023 314.904 10.9141 314.904 10.3555V9.05469C314.904 8.35547 314.967 7.76953 315.092 7.29688C315.221 6.82422 315.402 6.44531 315.637 6.16016C315.871 5.87109 316.152 5.66406 316.48 5.53906C316.812 5.41406 317.184 5.35156 317.594 5.35156C317.926 5.35156 318.23 5.39258 318.508 5.47461C318.789 5.55273 319.039 5.67969 319.258 5.85547C319.477 6.02734 319.662 6.25781 319.814 6.54688C319.971 6.83203 320.09 7.18164 320.172 7.5957C320.254 8.00977 320.295 8.49609 320.295 9.05469ZM319.205 10.5312V8.87305C319.205 8.49023 319.182 8.1543 319.135 7.86523C319.092 7.57227 319.027 7.32227 318.941 7.11523C318.855 6.9082 318.746 6.74023 318.613 6.61133C318.484 6.48242 318.334 6.38867 318.162 6.33008C317.994 6.26758 317.805 6.23633 317.594 6.23633C317.336 6.23633 317.107 6.28516 316.908 6.38281C316.709 6.47656 316.541 6.62695 316.404 6.83398C316.271 7.04102 316.17 7.3125 316.1 7.64844C316.029 7.98438 315.994 8.39258 315.994 8.87305V10.5312C315.994 10.9141 316.016 11.252 316.059 11.5449C316.105 11.8379 316.174 12.0918 316.264 12.3066C316.354 12.5176 316.463 12.6914 316.592 12.8281C316.721 12.9648 316.869 13.0664 317.037 13.1328C317.209 13.1953 317.398 13.2266 317.605 13.2266C317.871 13.2266 318.104 13.1758 318.303 13.0742C318.502 12.9727 318.668 12.8145 318.801 12.5996C318.938 12.3809 319.039 12.1016 319.105 11.7617C319.172 11.418 319.205 11.0078 319.205 10.5312ZM327.045 9.05469V10.3555C327.045 11.0547 326.982 11.6445 326.857 12.125C326.732 12.6055 326.553 12.9922 326.318 13.2852C326.084 13.5781 325.801 13.791 325.469 13.9238C325.141 14.0527 324.77 14.1172 324.355 14.1172C324.027 14.1172 323.725 14.0762 323.447 13.9941C323.17 13.9121 322.92 13.7812 322.697 13.6016C322.479 13.418 322.291 13.1797 322.135 12.8867C321.979 12.5938 321.859 12.2383 321.777 11.8203C321.695 11.4023 321.654 10.9141 321.654 10.3555V9.05469C321.654 8.35547 321.717 7.76953 321.842 7.29688C321.971 6.82422 322.152 6.44531 322.387 6.16016C322.621 5.87109 322.902 5.66406 323.23 5.53906C323.562 5.41406 323.934 5.35156 324.344 5.35156C324.676 5.35156 324.98 5.39258 325.258 5.47461C325.539 5.55273 325.789 5.67969 326.008 5.85547C326.227 6.02734 326.412 6.25781 326.564 6.54688C326.721 6.83203 326.84 7.18164 326.922 7.5957C327.004 8.00977 327.045 8.49609 327.045 9.05469ZM325.955 10.5312V8.87305C325.955 8.49023 325.932 8.1543 325.885 7.86523C325.842 7.57227 325.777 7.32227 325.691 7.11523C325.605 6.9082 325.496 6.74023 325.363 6.61133C325.234 6.48242 325.084 6.38867 324.912 6.33008C324.744 6.26758 324.555 6.23633 324.344 6.23633C324.086 6.23633 323.857 6.28516 323.658 6.38281C323.459 6.47656 323.291 6.62695 323.154 6.83398C323.021 7.04102 322.92 7.3125 322.85 7.64844C322.779 7.98438 322.744 8.39258 322.744 8.87305V10.5312C322.744 10.9141 322.766 11.252 322.809 11.5449C322.855 11.8379 322.924 12.0918 323.014 12.3066C323.104 12.5176 323.213 12.6914 323.342 12.8281C323.471 12.9648 323.619 13.0664 323.787 13.1328C323.959 13.1953 324.148 13.2266 324.355 13.2266C324.621 13.2266 324.854 13.1758 325.053 13.0742C325.252 12.9727 325.418 12.8145 325.551 12.5996C325.688 12.3809 325.789 12.1016 325.855 11.7617C325.922 11.418 325.955 11.0078 325.955 10.5312ZM328.346 7.56055V7.10938C328.346 6.78516 328.416 6.49023 328.557 6.22461C328.697 5.95898 328.898 5.74609 329.16 5.58594C329.422 5.42578 329.732 5.3457 330.092 5.3457C330.459 5.3457 330.771 5.42578 331.029 5.58594C331.291 5.74609 331.492 5.95898 331.633 6.22461C331.773 6.49023 331.844 6.78516 331.844 7.10938V7.56055C331.844 7.87695 331.773 8.16797 331.633 8.43359C331.496 8.69922 331.297 8.91211 331.035 9.07227C330.777 9.23242 330.467 9.3125 330.104 9.3125C329.74 9.3125 329.426 9.23242 329.16 9.07227C328.898 8.91211 328.697 8.69922 328.557 8.43359C328.416 8.16797 328.346 7.87695 328.346 7.56055ZM329.16 7.10938V7.56055C329.16 7.74023 329.193 7.91016 329.26 8.07031C329.33 8.23047 329.436 8.36133 329.576 8.46289C329.717 8.56055 329.893 8.60938 330.104 8.60938C330.314 8.60938 330.488 8.56055 330.625 8.46289C330.762 8.36133 330.863 8.23047 330.93 8.07031C330.996 7.91016 331.029 7.74023 331.029 7.56055V7.10938C331.029 6.92578 330.994 6.75391 330.924 6.59375C330.857 6.42969 330.754 6.29883 330.613 6.20117C330.477 6.09961 330.303 6.04883 330.092 6.04883C329.885 6.04883 329.711 6.09961 329.57 6.20117C329.434 6.29883 329.33 6.42969 329.26 6.59375C329.193 6.75391 329.16 6.92578 329.16 7.10938ZM332.5 12.3652V11.9082C332.5 11.5879 332.57 11.2949 332.711 11.0293C332.852 10.7637 333.053 10.5508 333.314 10.3906C333.576 10.2305 333.887 10.1504 334.246 10.1504C334.613 10.1504 334.926 10.2305 335.184 10.3906C335.445 10.5508 335.646 10.7637 335.787 11.0293C335.928 11.2949 335.998 11.5879 335.998 11.9082V12.3652C335.998 12.6855 335.928 12.9785 335.787 13.2441C335.65 13.5098 335.451 13.7227 335.189 13.8828C334.932 14.043 334.621 14.123 334.258 14.123C333.895 14.123 333.582 14.043 333.32 13.8828C333.059 13.7227 332.855 13.5098 332.711 13.2441C332.57 12.9785 332.5 12.6855 332.5 12.3652ZM333.314 11.9082V12.3652C333.314 12.5449 333.348 12.7168 333.414 12.8809C333.484 13.041 333.59 13.1719 333.73 13.2734C333.871 13.3711 334.047 13.4199 334.258 13.4199C334.469 13.4199 334.643 13.3711 334.779 13.2734C334.92 13.1719 335.023 13.041 335.09 12.8809C335.156 12.7207 335.189 12.5488 335.189 12.3652V11.9082C335.189 11.7246 335.154 11.5527 335.084 11.3926C335.018 11.2324 334.914 11.1035 334.773 11.0059C334.637 10.9043 334.461 10.8535 334.246 10.8535C334.039 10.8535 333.865 10.9043 333.725 11.0059C333.588 11.1035 333.484 11.2324 333.414 11.3926C333.348 11.5527 333.314 11.7246 333.314 11.9082ZM334.516 6.68164L330.35 13.3496L329.74 12.9629L333.906 6.29492L334.516 6.68164Z" fill="#111111" />
-      <rect x="345.036" y="5.15385" width="21.3818" height="9.69231" rx="1.33" fill="#111111" />
-      <path opacity="0.4" fill-rule="evenodd" clip-rule="evenodd" d="M348.6 4H362.855C363.851 4 364.538 4.00078 365.072 4.04436C365.594 4.087 365.88 4.16555 366.09 4.27248C366.56 4.51217 366.942 4.89462 367.182 5.36502C367.289 5.5749 367.368 5.86085 367.41 6.38264C367.454 6.9161 367.455 7.60341 367.455 8.6V11.4C367.455 12.3966 367.454 13.0839 367.41 13.6174C367.368 14.1392 367.289 14.4251 367.182 14.635C366.942 15.1054 366.56 15.4878 366.09 15.7275C365.88 15.8345 365.594 15.913 365.072 15.9556C364.538 15.9992 363.851 16 362.855 16H348.6C347.603 16 346.916 15.9992 346.383 15.9556C345.861 15.913 345.575 15.8345 345.365 15.7275C344.895 15.4878 344.512 15.1054 344.272 14.635C344.166 14.4251 344.087 14.1392 344.044 13.6174C344.001 13.0839 344 12.3966 344 11.4V8.6C344 7.60341 344.001 6.9161 344.044 6.38264C344.087 5.86085 344.166 5.5749 344.272 5.36502C344.512 4.89462 344.895 4.51217 345.365 4.27248C345.575 4.16555 345.861 4.087 346.383 4.04436C346.916 4.00078 347.603 4 348.6 4ZM343 8.6C343 6.63982 343 5.65972 343.381 4.91103C343.717 4.25247 344.252 3.71703 344.911 3.38148C345.66 3 346.64 3 348.6 3H362.855C364.815 3 365.795 3 366.544 3.38148C367.202 3.71703 367.738 4.25247 368.073 4.91103C368.455 5.65972 368.455 6.63982 368.455 8.6V11.4C368.455 13.3602 368.455 14.3403 368.073 15.089C367.738 15.7475 367.202 16.283 366.544 16.6185C365.795 17 364.815 17 362.855 17H348.6C346.64 17 345.66 17 344.911 16.6185C344.252 16.283 343.717 15.7475 343.381 15.089C343 14.3403 343 13.3602 343 11.4V8.6ZM369.473 12.1538C370.316 11.9385 371 11.1895 371 10C371 8.81046 370.316 8.06154 369.473 7.84615L369.473 10L369.473 12.1538Z" fill="#111111" />
-      <defs>
-        <clipPath id="clip0_288_439">
-          <rect width="20" height="14" fill="white" transform="translate(4 3)" />
-        </clipPath>
-      </defs>
-    </svg>
-  )
-};
-
-export default StatusBar;
