--- src/frames/iphone-6s-plus/svg/device-frame.tsx
@@ -1,80 +1,0 @@
-import { concat } from "nixix/primitives";
-
-const DeviceFrame: Nixix.FC<App.SVGProps> = ({ className, ...rest }): someView => {
-  const resizeClass = ' '
-
-  return (
-    <svg
-      {...rest} className={concat` ${resizeClass} ${className} tws-w-[auto] tws-h-[736px] `} viewBox="0 0 474 952" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path d="M8.76921 109.654L8.76921 109.103L8.21792 109.103L2.7051 109.103C1.79171 109.103 1.05126 109.843 1.05126 110.757L1.05126 134.426C1.05126 138.099 4.02914 141.077 7.70254 141.077L8.21793 141.077L8.76921 141.077L8.76921 140.526L8.76921 109.654Z" fill="black" stroke="url(#paint0_linear_296_491)" stroke-width="1.10256" />
-      <path d="M8.76921 182.423L8.76921 181.872L8.21792 181.872L2.7051 181.872C1.7917 181.872 1.05125 182.613 1.05125 183.526L1.05126 232.554C1.05126 236.228 4.02914 239.205 7.70254 239.205L8.21793 239.205L8.76921 239.205L8.76921 238.654L8.76921 182.423Z" fill="black" stroke="url(#paint1_linear_296_491)" stroke-width="1.10256" />
-      <path d="M8.76921 257.398L8.76921 256.846L8.21792 256.846L2.7051 256.846C1.7917 256.846 1.05125 257.587 1.05125 258.5L1.05126 307.529C1.05126 311.202 4.02914 314.18 7.70254 314.18L8.21793 314.18L8.76921 314.18L8.76921 313.629L8.76921 257.398Z" fill="black" stroke="url(#paint2_linear_296_491)" stroke-width="1.10256" />
-      <path d="M466.333 238.654L466.333 239.205L466.885 239.205L471.295 239.205C472.208 239.205 472.949 238.465 472.949 237.552L472.949 185.731C472.949 182.382 470.234 179.667 466.885 179.667L466.885 180.218L466.333 180.218L466.333 238.654Z" fill="black" stroke="url(#paint3_linear_296_491)" stroke-width="1.10256" />
-      <path d="M467.797 888.565C467.727 922.855 439.595 950.91 405.285 950.91H66.6539C32.3438 950.91 4.27197 922.855 4.27197 888.565V62.8446C4.27197 28.555 32.3438 0.499878 66.6539 0.499878H407.011C441.321 0.499878 469.333 28.555 469.263 62.8446L467.797 888.565Z" fill="#595A60" />
-      <path d="M66.6538 948.139C33.7841 948.139 7.04443 921.414 7.04443 888.565V62.8442C7.04443 29.9954 33.7841 3.27032 66.6538 3.27032H407.011C422.864 3.27032 437.798 9.4854 449.067 20.7698C460.336 32.0542 466.523 46.9947 466.49 62.8386L465.025 888.56C464.957 921.411 438.158 948.139 405.285 948.139H66.6538Z" fill="#83848A" />
-      <path d="M66.6542 945.369C35.315 945.369 9.81738 919.886 9.81738 888.566V62.8448C9.81738 31.5242 35.315 6.04181 66.6542 6.04181H407.011C422.122 6.04181 436.361 11.9673 447.105 22.728C457.848 33.4887 463.751 47.731 463.718 62.8337L462.253 888.554C462.187 919.882 436.633 945.369 405.285 945.369H66.6542Z" fill="#111111" />
-      <path d="M66.6542 942.597C36.8412 942.597 12.5898 918.359 12.5898 888.565V62.8443C12.5898 33.0504 36.8412 8.81219 66.6542 8.81219H407.011C421.38 8.81219 434.921 14.4495 445.142 24.6852C455.363 34.9222 460.976 48.469 460.945 62.8276L459.48 888.548C459.418 918.351 435.107 942.597 405.285 942.597H66.6542Z" fill="black" />
-      <path d="M466.488 63.2832L466.465 75.2132H469.237L469.261 63.2832H466.488Z" fill="#505050" />
-      <path d="M463.692 75.2132H466.465L466.488 63.2832H463.716L463.692 75.2132Z" fill="#6E6E6E" />
-      <path d="M7.0445 63.2832L7.06807 75.2132H4.29554L4.27197 63.2832H7.0445Z" fill="#505050" />
-      <path d="M9.84053 75.2132H7.068L7.04443 63.2832H9.81696L9.84053 75.2132Z" fill="#6E6E6E" />
-      <circle cx="237" cy="894.128" r="31.4231" stroke="url(#paint4_angular_296_491)" stroke-opacity="0.3" stroke-width="4.41026" />
-      <circle cx="237" cy="894.128" r="29.2179" fill="url(#paint5_angular_296_491)" fill-opacity="0.16" />
-      <circle cx="237" cy="894.128" r="28.9974" stroke="white" stroke-opacity="0.4" stroke-width="0.441026" />
-      <path fill-rule="evenodd" clip-rule="evenodd" d="M180.072 59.5996C180.07 61.4801 179.323 63.2831 177.992 64.6123C176.662 65.9416 174.859 66.6883 172.978 66.6883C171.098 66.6883 169.294 65.9416 167.964 64.6123C166.634 63.2831 165.886 61.4801 165.885 59.5996C165.884 58.6677 166.067 57.7448 166.423 56.8836C166.78 56.0224 167.302 55.2398 167.961 54.5806C168.619 53.9214 169.402 53.3985 170.263 53.0417C171.124 52.6849 172.046 52.5013 172.978 52.5013C173.91 52.5013 174.833 52.6849 175.694 53.0417C176.555 53.3985 177.337 53.9214 177.996 54.5806C178.655 55.2398 179.177 56.0224 179.533 56.8836C179.889 57.7448 180.072 58.6677 180.072 59.5996Z" fill="url(#paint6_linear_296_491)" />
-      <path d="M177.525 59.8157C177.524 61.0602 177.029 62.2534 176.149 63.1331C175.269 64.0128 174.075 64.507 172.831 64.507C171.586 64.507 170.393 64.0128 169.513 63.1331C168.632 62.2534 168.137 61.0602 168.137 59.8157C168.136 59.199 168.257 58.5883 168.493 58.0184C168.729 57.4485 169.075 56.9307 169.511 56.4945C169.947 56.0583 170.464 55.7122 171.034 55.4761C171.604 55.24 172.214 55.1185 172.831 55.1185C173.448 55.1185 174.058 55.24 174.628 55.4761C175.198 55.7122 175.715 56.0583 176.151 56.4945C176.587 56.9307 176.933 57.4485 177.169 58.0184C177.404 58.5883 177.526 59.199 177.525 59.8157Z" fill="black" />
-      <path d="M177.217 59.815C177.217 60.9784 176.755 62.0941 175.932 62.9167C175.11 63.7393 173.994 64.2015 172.831 64.2015C171.667 64.2015 170.552 63.7393 169.729 62.9167C168.906 62.0941 168.444 60.9784 168.444 59.815C168.444 58.6517 168.906 57.536 169.729 56.7134C170.552 55.8908 171.667 55.4286 172.831 55.4286C173.994 55.4286 175.11 55.8908 175.932 56.7134C176.755 57.536 177.217 58.6517 177.217 59.815Z" fill="#001A49" />
-      <path opacity="0.75" d="M175.542 59.6619C175.54 60.3593 175.262 61.0275 174.768 61.5199C174.274 62.0124 173.605 62.289 172.908 62.289C172.21 62.289 171.542 62.0124 171.048 61.5199C170.554 61.0275 170.276 60.3593 170.274 59.6619C170.273 59.3154 170.341 58.9722 170.472 58.6518C170.604 58.3315 170.798 58.0403 171.043 57.795C171.288 57.5497 171.578 57.3551 171.898 57.2223C172.218 57.0895 172.561 57.0211 172.908 57.0211C173.254 57.0211 173.597 57.0895 173.917 57.2223C174.237 57.3551 174.528 57.5497 174.773 57.795C175.017 58.0403 175.211 58.3315 175.343 58.6518C175.475 58.9722 175.543 59.3154 175.542 59.6619Z" fill="#232323" stroke="#555555" stroke-width="0.633731" />
-      <g filter="url(#filter0_f_296_491)">
-        <path d="M173.235 56.7328C171.877 56.6244 169.792 58.4051 169.419 60.3615C170.902 59.8796 172.513 57.484 173.235 56.7328Z" fill="white" />
-      </g>
-      <path d="M268.464 56.1489H205.443C203.639 56.1489 202.176 57.6115 202.176 59.4157C202.176 61.2199 203.639 62.6825 205.443 62.6825H268.464C270.268 62.6825 271.731 61.2199 271.731 59.4157C271.731 57.6115 270.268 56.1489 268.464 56.1489Z" fill="#232426" />
-      <path d="M236.953 38.0911C239.418 38.0911 241.416 36.0932 241.416 33.6287C241.416 31.1641 239.418 29.1662 236.953 29.1662C234.489 29.1662 232.491 31.1641 232.491 33.6287C232.491 36.0932 234.489 38.0911 236.953 38.0911Z" fill="#2C2C2C" />
-      <rect width="414" height="736" transform="translate(29.5 107.5)" fill="white" />
-      <defs>
-        <filter id="filter0_f_296_491" x="167.708" y="55.0167" width="7.23812" height="7.05617" filterUnits="userSpaceOnUse" color-interpolation-filters="sRGB">
-          <feFlood flood-opacity="0" result="BackgroundImageFix" />
-          <feBlend mode="normal" in="SourceGraphic" in2="BackgroundImageFix" result="shape" />
-          <feGaussianBlur stdDeviation="0.855671" result="effect1_foregroundBlur_296_491" />
-        </filter>
-        <linearGradient id="paint0_linear_296_491" x1="1.60254" y1="125.09" x2="8.21792" y2="125.09" gradientUnits="userSpaceOnUse">
-          <stop stop-color="#474643" />
-          <stop offset="1" stop-color="#5E5E5E" />
-        </linearGradient>
-        <linearGradient id="paint1_linear_296_491" x1="1.60254" y1="210.539" x2="8.21793" y2="210.539" gradientUnits="userSpaceOnUse">
-          <stop stop-color="#474643" />
-          <stop offset="1" stop-color="#5E5E5E" />
-        </linearGradient>
-        <linearGradient id="paint2_linear_296_491" x1="1.60254" y1="285.513" x2="8.21793" y2="285.513" gradientUnits="userSpaceOnUse">
-          <stop stop-color="#474643" />
-          <stop offset="1" stop-color="#5E5E5E" />
-        </linearGradient>
-        <linearGradient id="paint3_linear_296_491" x1="472.397" y1="209.436" x2="466.885" y2="209.436" gradientUnits="userSpaceOnUse">
-          <stop stop-color="#474643" />
-          <stop offset="1" stop-color="#5E5E5E" />
-        </linearGradient>
-        <radialGradient id="paint4_angular_296_491" cx="0" cy="0" r="1" gradientUnits="userSpaceOnUse" gradientTransform="translate(236.124 894.128) rotate(90) scale(29.218 28.3414)">
-          <stop offset="0.208333" stop-color="#D6DFE6" />
-          <stop offset="0.370506" stop-color="#FAFBFC" stop-opacity="0.53" />
-          <stop offset="0.755208" stop-color="#D7E0E7" />
-          <stop offset="0.911458" stop-color="#F7F8F9" stop-opacity="0.19" />
-        </radialGradient>
-        <radialGradient id="paint5_angular_296_491" cx="0" cy="0" r="1" gradientUnits="userSpaceOnUse" gradientTransform="translate(236.124 894.128) rotate(90) scale(29.218 28.3414)">
-          <stop offset="0.0729167" stop-color="#5A5A5A" />
-          <stop offset="0.244792" />
-          <stop offset="0.40625" stop-color="#3A3A3A" />
-          <stop offset="0.567708" />
-          <stop offset="0.744792" stop-color="#5E5E5E" />
-          <stop offset="0.921875" />
-        </radialGradient>
-        <linearGradient id="paint6_linear_296_491" x1="167.911" y1="54.6357" x2="178.267" y2="64.3266" gradientUnits="userSpaceOnUse">
-          <stop stop-color="#1C1518" />
-          <stop offset="1" stop-color="#191715" />
-        </linearGradient>
-      </defs>
-    </svg>
-
-  )
-}
-
-export default DeviceFrame;
