--- src/frames/iphone-6s-plus/svg/status-bar.tsx
@@ -1,25 +1,0 @@
-const StatusBar: Nixix.FC<App.SVGProps> = ({ className, ...rest }): someView => {
-  return (
-    <svg width="414" height="20" viewBox="0 0 414 20" fill="none" xmlns="http://www.w3.org/2000/svg"
-      {...rest} className={className + ' tws-h-fit '}>
-      <g clip-path="url(#clip0_296_523)">
-        <path fill-rule="evenodd" clip-rule="evenodd" d="M21.6492 4C21.097 4 20.6492 4.44772 20.6492 5V15C20.6492 15.5523 21.0969 16 21.6492 16H23C23.5523 16 24 15.5523 24 15V5C24 4.44772 23.5523 4 23 4H21.6492ZM16.0995 6.40002C15.5472 6.40002 15.0995 6.84774 15.0995 7.40002V15C15.0995 15.5523 15.5472 16 16.0995 16H17.4503C18.0026 16 18.4503 15.5523 18.4503 15V7.40002C18.4503 6.84774 18.0026 6.40002 17.4503 6.40002H16.0995ZM9.54983 10C9.54983 9.44772 9.99755 9 10.5498 9H11.9006C12.4529 9 12.9006 9.44772 12.9006 10V15C12.9006 15.5523 12.4529 16 11.9006 16H10.5498C9.99754 16 9.54983 15.5523 9.54983 15V10ZM5 11.4C4.44772 11.4 4 11.8477 4 12.4V15C4 15.5523 4.44771 16 5 16H6.35078C6.90307 16 7.35078 15.5523 7.35078 15V12.4C7.35078 11.8477 6.90307 11.4 6.35078 11.4H5Z" fill="#111111" />
-      </g>
-      <path d="M36.2285 11.2871H37.3535C37.2949 11.8262 37.1406 12.3086 36.8906 12.7344C36.6406 13.1602 36.2871 13.498 35.8301 13.748C35.373 13.9941 34.8027 14.1172 34.1191 14.1172C33.6191 14.1172 33.1641 14.0234 32.7539 13.8359C32.3477 13.6484 31.998 13.3828 31.7051 13.0391C31.4121 12.6914 31.1855 12.2754 31.0254 11.791C30.8691 11.3027 30.791 10.7598 30.791 10.1621V9.3125C30.791 8.71484 30.8691 8.17383 31.0254 7.68945C31.1855 7.20117 31.4141 6.7832 31.7109 6.43555C32.0117 6.08789 32.373 5.82031 32.7949 5.63281C33.2168 5.44531 33.6914 5.35156 34.2188 5.35156C34.8633 5.35156 35.4082 5.47266 35.8535 5.71484C36.2988 5.95703 36.6445 6.29297 36.8906 6.72266C37.1406 7.14844 37.2949 7.64258 37.3535 8.20508H36.2285C36.1738 7.80664 36.0723 7.46484 35.9238 7.17969C35.7754 6.89062 35.5645 6.66797 35.291 6.51172C35.0176 6.35547 34.6602 6.27734 34.2188 6.27734C33.8398 6.27734 33.5059 6.34961 33.2168 6.49414C32.9316 6.63867 32.6914 6.84375 32.4961 7.10938C32.3047 7.375 32.1602 7.69336 32.0625 8.06445C31.9648 8.43555 31.916 8.84766 31.916 9.30078V10.1621C31.916 10.5801 31.959 10.9727 32.0449 11.3398C32.1348 11.707 32.2695 12.0293 32.4492 12.3066C32.6289 12.584 32.8574 12.8027 33.1348 12.9629C33.4121 13.1191 33.7402 13.1973 34.1191 13.1973C34.5996 13.1973 34.9824 13.1211 35.2676 12.9688C35.5527 12.8164 35.7676 12.5977 35.9121 12.3125C36.0605 12.0273 36.166 11.6855 36.2285 11.2871ZM42.4688 12.916V9.65234C42.4688 9.40234 42.418 9.18555 42.3164 9.00195C42.2188 8.81445 42.0703 8.66992 41.8711 8.56836C41.6719 8.4668 41.4258 8.41602 41.1328 8.41602C40.8594 8.41602 40.6191 8.46289 40.4121 8.55664C40.209 8.65039 40.0488 8.77344 39.9316 8.92578C39.8184 9.07812 39.7617 9.24219 39.7617 9.41797H38.6777C38.6777 9.19141 38.7363 8.9668 38.8535 8.74414C38.9707 8.52148 39.1387 8.32031 39.3574 8.14062C39.5801 7.95703 39.8457 7.8125 40.1543 7.70703C40.4668 7.59766 40.8145 7.54297 41.1973 7.54297C41.6582 7.54297 42.0645 7.62109 42.416 7.77734C42.7715 7.93359 43.0488 8.16992 43.248 8.48633C43.4512 8.79883 43.5527 9.19141 43.5527 9.66406V12.6172C43.5527 12.8281 43.5703 13.0527 43.6055 13.291C43.6445 13.5293 43.7012 13.7344 43.7754 13.9062V14H42.6445C42.5898 13.875 42.5469 13.709 42.5156 13.502C42.4844 13.291 42.4688 13.0957 42.4688 12.916ZM42.6562 10.1562L42.668 10.918H41.5723C41.2637 10.918 40.9883 10.9434 40.7461 10.9941C40.5039 11.041 40.3008 11.1133 40.1367 11.2109C39.9727 11.3086 39.8477 11.4316 39.7617 11.5801C39.6758 11.7246 39.6328 11.8945 39.6328 12.0898C39.6328 12.2891 39.6777 12.4707 39.7676 12.6348C39.8574 12.7988 39.9922 12.9297 40.1719 13.0273C40.3555 13.1211 40.5801 13.168 40.8457 13.168C41.1777 13.168 41.4707 13.0977 41.7246 12.957C41.9785 12.8164 42.1797 12.6445 42.3281 12.4414C42.4805 12.2383 42.5625 12.041 42.5742 11.8496L43.0371 12.3711C43.0098 12.5352 42.9355 12.7168 42.8145 12.916C42.6934 13.1152 42.5312 13.3066 42.3281 13.4902C42.1289 13.6699 41.8906 13.8203 41.6133 13.9414C41.3398 14.0586 41.0312 14.1172 40.6875 14.1172C40.2578 14.1172 39.8809 14.0332 39.5566 13.8652C39.2363 13.6973 38.9863 13.4727 38.8066 13.1914C38.6309 12.9062 38.543 12.5879 38.543 12.2363C38.543 11.8965 38.6094 11.5977 38.7422 11.3398C38.875 11.0781 39.0664 10.8613 39.3164 10.6895C39.5664 10.5137 39.8672 10.3809 40.2188 10.291C40.5703 10.2012 40.9629 10.1562 41.3965 10.1562H42.6562ZM46.3418 8.65625V14H45.2578V7.66016H46.3125L46.3418 8.65625ZM48.3223 7.625L48.3164 8.63281C48.2266 8.61328 48.1406 8.60156 48.0586 8.59766C47.9805 8.58984 47.8906 8.58594 47.7891 8.58594C47.5391 8.58594 47.3184 8.625 47.127 8.70312C46.9355 8.78125 46.7734 8.89062 46.6406 9.03125C46.5078 9.17188 46.4023 9.33984 46.3242 9.53516C46.25 9.72656 46.2012 9.9375 46.1777 10.168L45.873 10.3438C45.873 9.96094 45.9102 9.60156 45.9844 9.26562C46.0625 8.92969 46.1816 8.63281 46.3418 8.375C46.502 8.11328 46.7051 7.91016 46.9512 7.76562C47.2012 7.61719 47.498 7.54297 47.8418 7.54297C47.9199 7.54297 48.0098 7.55273 48.1113 7.57227C48.2129 7.58789 48.2832 7.60547 48.3223 7.625ZM50.4082 8.65625V14H49.3242V7.66016H50.3789L50.4082 8.65625ZM52.3887 7.625L52.3828 8.63281C52.293 8.61328 52.207 8.60156 52.125 8.59766C52.0469 8.58984 51.957 8.58594 51.8555 8.58594C51.6055 8.58594 51.3848 8.625 51.1934 8.70312C51.002 8.78125 50.8398 8.89062 50.707 9.03125C50.5742 9.17188 50.4688 9.33984 50.3906 9.53516C50.3164 9.72656 50.2676 9.9375 50.2441 10.168L49.9395 10.3438C49.9395 9.96094 49.9766 9.60156 50.0508 9.26562C50.1289 8.92969 50.248 8.63281 50.4082 8.375C50.5684 8.11328 50.7715 7.91016 51.0176 7.76562C51.2676 7.61719 51.5645 7.54297 51.9082 7.54297C51.9863 7.54297 52.0762 7.55273 52.1777 7.57227C52.2793 7.58789 52.3496 7.60547 52.3887 7.625ZM54.5684 7.66016V14H53.4785V7.66016H54.5684ZM53.3965 5.97852C53.3965 5.80273 53.4492 5.6543 53.5547 5.5332C53.6641 5.41211 53.8242 5.35156 54.0352 5.35156C54.2422 5.35156 54.4004 5.41211 54.5098 5.5332C54.623 5.6543 54.6797 5.80273 54.6797 5.97852C54.6797 6.14648 54.623 6.29102 54.5098 6.41211C54.4004 6.5293 54.2422 6.58789 54.0352 6.58789C53.8242 6.58789 53.6641 6.5293 53.5547 6.41211C53.4492 6.29102 53.3965 6.14648 53.3965 5.97852ZM58.9395 14.1172C58.498 14.1172 58.0977 14.043 57.7383 13.8945C57.3828 13.7422 57.0762 13.5293 56.8184 13.2559C56.5645 12.9824 56.3691 12.6582 56.2324 12.2832C56.0957 11.9082 56.0273 11.498 56.0273 11.0527V10.8066C56.0273 10.291 56.1035 9.83203 56.2559 9.42969C56.4082 9.02344 56.6152 8.67969 56.877 8.39844C57.1387 8.11719 57.4355 7.9043 57.7676 7.75977C58.0996 7.61523 58.4434 7.54297 58.7988 7.54297C59.252 7.54297 59.6426 7.62109 59.9707 7.77734C60.3027 7.93359 60.5742 8.15234 60.7852 8.43359C60.9961 8.71094 61.1523 9.03906 61.2539 9.41797C61.3555 9.79297 61.4062 10.2031 61.4062 10.6484V11.1348H56.6719V10.25H60.3223V10.168C60.3066 9.88672 60.248 9.61328 60.1465 9.34766C60.0488 9.08203 59.8926 8.86328 59.6777 8.69141C59.4629 8.51953 59.1699 8.43359 58.7988 8.43359C58.5527 8.43359 58.3262 8.48633 58.1191 8.5918C57.9121 8.69336 57.7344 8.8457 57.5859 9.04883C57.4375 9.25195 57.3223 9.5 57.2402 9.79297C57.1582 10.0859 57.1172 10.4238 57.1172 10.8066V11.0527C57.1172 11.3535 57.1582 11.6367 57.2402 11.9023C57.3262 12.1641 57.4492 12.3945 57.6094 12.5938C57.7734 12.793 57.9707 12.9492 58.2012 13.0625C58.4355 13.1758 58.7012 13.2324 58.998 13.2324C59.3809 13.2324 59.7051 13.1543 59.9707 12.998C60.2363 12.8418 60.4688 12.6328 60.668 12.3711L61.3242 12.8926C61.1875 13.0996 61.0137 13.2969 60.8027 13.4844C60.5918 13.6719 60.332 13.8242 60.0234 13.9414C59.7188 14.0586 59.3574 14.1172 58.9395 14.1172ZM63.7559 8.65625V14H62.6719V7.66016H63.7266L63.7559 8.65625ZM65.7363 7.625L65.7305 8.63281C65.6406 8.61328 65.5547 8.60156 65.4727 8.59766C65.3945 8.58984 65.3047 8.58594 65.2031 8.58594C64.9531 8.58594 64.7324 8.625 64.541 8.70312C64.3496 8.78125 64.1875 8.89062 64.0547 9.03125C63.9219 9.17188 63.8164 9.33984 63.7383 9.53516C63.6641 9.72656 63.6152 9.9375 63.5918 10.168L63.2871 10.3438C63.2871 9.96094 63.3242 9.60156 63.3984 9.26562C63.4766 8.92969 63.5957 8.63281 63.7559 8.375C63.916 8.11328 64.1191 7.91016 64.3652 7.76562C64.6152 7.61719 64.9121 7.54297 65.2559 7.54297C65.334 7.54297 65.4238 7.55273 65.5254 7.57227C65.627 7.58789 65.6973 7.60547 65.7363 7.625Z" fill="#111111" />
-      <path d="M82.0089 12.3275C82.7269 12.3275 83.4246 12.5292 84.0394 12.9137L84.2885 13.0695C84.4773 13.1876 84.5105 13.464 84.3557 13.629L82.229 15.8956C82.0984 16.0348 81.8866 16.0348 81.7559 15.8956L79.6429 13.6436C79.4888 13.4793 79.5209 13.2043 79.7081 13.0853L79.9544 12.9288C80.5748 12.5346 81.2814 12.3275 82.0089 12.3275Z" fill="#111111" />
-      <path d="M82.0086 8.16388C83.7948 8.16388 85.5098 8.79717 86.9069 9.96996L87.1044 10.1357C87.2618 10.2678 87.2756 10.5168 87.1339 10.6678L85.8643 12.021C85.7469 12.1462 85.561 12.1605 85.4279 12.0548L85.2731 11.9318C84.3295 11.1823 83.1917 10.7803 82.0086 10.7803C80.8183 10.7803 79.6738 11.1873 78.7271 11.9454L78.5722 12.0695C78.439 12.1761 78.2525 12.1621 78.1347 12.0366L76.8656 10.6839C76.7242 10.5332 76.7376 10.2848 76.8943 10.1525L77.0906 9.98658C78.491 8.80334 80.2138 8.16388 82.0086 8.16388Z" fill="#111111" />
-      <path d="M82.0087 4C84.8434 4 87.5542 5.06869 89.7 7.02754L89.8831 7.19474C90.0311 7.32986 90.0398 7.5707 89.902 7.71757L88.6365 9.06634C88.5139 9.19702 88.3179 9.20617 88.1851 9.08744L88.0284 8.9474C86.339 7.43766 84.2215 6.61643 82.0087 6.61643C79.7883 6.61643 77.664 7.44332 75.9721 8.9625L75.8154 9.10325C75.6825 9.22252 75.4861 9.21359 75.3633 9.0827L74.098 7.73412C73.9604 7.58745 73.9688 7.34702 74.1164 7.21176L74.2988 7.0445C76.4472 5.07507 79.1655 4 82.0087 4Z" fill="#111111" />
-      <path d="M185.502 5.44531V14H183.949V7.25586L181.893 7.93555V6.68164L185.326 5.44531H185.502ZM194.021 9.00195V10.4375C194.021 11.0938 193.955 11.6543 193.822 12.1191C193.689 12.584 193.498 12.9648 193.248 13.2617C192.998 13.5547 192.699 13.7715 192.352 13.9121C192.004 14.0488 191.617 14.1172 191.191 14.1172C190.852 14.1172 190.535 14.0742 190.242 13.9883C189.949 13.8984 189.686 13.7617 189.451 13.5781C189.221 13.3906 189.021 13.1523 188.854 12.8633C188.689 12.5742 188.562 12.2285 188.473 11.8262C188.387 11.4238 188.344 10.9609 188.344 10.4375V9.00195C188.344 8.3457 188.41 7.78711 188.543 7.32617C188.68 6.86523 188.873 6.48828 189.123 6.19531C189.373 5.90234 189.672 5.68945 190.02 5.55664C190.367 5.41992 190.754 5.35156 191.18 5.35156C191.523 5.35156 191.84 5.39453 192.129 5.48047C192.422 5.56641 192.684 5.70117 192.914 5.88477C193.148 6.06445 193.348 6.29883 193.512 6.58789C193.676 6.87305 193.801 7.2168 193.887 7.61914C193.977 8.01758 194.021 8.47852 194.021 9.00195ZM192.475 10.6484V8.78516C192.475 8.45703 192.455 8.17188 192.416 7.92969C192.381 7.68359 192.328 7.47461 192.258 7.30273C192.188 7.13086 192.1 6.99219 191.994 6.88672C191.889 6.77734 191.768 6.69727 191.631 6.64648C191.494 6.5957 191.344 6.57031 191.18 6.57031C190.977 6.57031 190.795 6.61133 190.635 6.69336C190.475 6.77148 190.34 6.89648 190.23 7.06836C190.121 7.24023 190.037 7.46875 189.979 7.75391C189.924 8.03516 189.896 8.37891 189.896 8.78516V10.6484C189.896 10.9766 189.914 11.2637 189.949 11.5098C189.988 11.7559 190.043 11.9668 190.113 12.1426C190.188 12.3184 190.275 12.4629 190.377 12.5762C190.482 12.6855 190.604 12.7676 190.74 12.8223C190.877 12.873 191.027 12.8984 191.191 12.8984C191.398 12.8984 191.58 12.8574 191.736 12.7754C191.896 12.6934 192.031 12.5645 192.141 12.3887C192.25 12.209 192.332 11.9766 192.387 11.6914C192.445 11.4023 192.475 11.0547 192.475 10.6484ZM195.346 13.2441C195.346 13.0098 195.426 12.8125 195.586 12.6523C195.746 12.4922 195.963 12.4121 196.236 12.4121C196.51 12.4121 196.727 12.4922 196.887 12.6523C197.047 12.8125 197.127 13.0098 197.127 13.2441C197.127 13.4746 197.047 13.6699 196.887 13.8301C196.727 13.9863 196.51 14.0645 196.236 14.0645C195.963 14.0645 195.746 13.9863 195.586 13.8301C195.426 13.6699 195.346 13.4746 195.346 13.2441ZM195.352 8.32812C195.352 8.09375 195.432 7.89648 195.592 7.73633C195.752 7.57617 195.969 7.49609 196.242 7.49609C196.516 7.49609 196.732 7.57617 196.893 7.73633C197.053 7.89648 197.133 8.09375 197.133 8.32812C197.133 8.55859 197.053 8.75391 196.893 8.91406C196.732 9.07031 196.516 9.14844 196.242 9.14844C195.969 9.14844 195.752 9.07031 195.592 8.91406C195.432 8.75391 195.352 8.55859 195.352 8.32812ZM204.146 9.00195V10.4375C204.146 11.0938 204.08 11.6543 203.947 12.1191C203.814 12.584 203.623 12.9648 203.373 13.2617C203.123 13.5547 202.824 13.7715 202.477 13.9121C202.129 14.0488 201.742 14.1172 201.316 14.1172C200.977 14.1172 200.66 14.0742 200.367 13.9883C200.074 13.8984 199.811 13.7617 199.576 13.5781C199.346 13.3906 199.146 13.1523 198.979 12.8633C198.814 12.5742 198.688 12.2285 198.598 11.8262C198.512 11.4238 198.469 10.9609 198.469 10.4375V9.00195C198.469 8.3457 198.535 7.78711 198.668 7.32617C198.805 6.86523 198.998 6.48828 199.248 6.19531C199.498 5.90234 199.797 5.68945 200.145 5.55664C200.492 5.41992 200.879 5.35156 201.305 5.35156C201.648 5.35156 201.965 5.39453 202.254 5.48047C202.547 5.56641 202.809 5.70117 203.039 5.88477C203.273 6.06445 203.473 6.29883 203.637 6.58789C203.801 6.87305 203.926 7.2168 204.012 7.61914C204.102 8.01758 204.146 8.47852 204.146 9.00195ZM202.6 10.6484V8.78516C202.6 8.45703 202.58 8.17188 202.541 7.92969C202.506 7.68359 202.453 7.47461 202.383 7.30273C202.312 7.13086 202.225 6.99219 202.119 6.88672C202.014 6.77734 201.893 6.69727 201.756 6.64648C201.619 6.5957 201.469 6.57031 201.305 6.57031C201.102 6.57031 200.92 6.61133 200.76 6.69336C200.6 6.77148 200.465 6.89648 200.355 7.06836C200.246 7.24023 200.162 7.46875 200.104 7.75391C200.049 8.03516 200.021 8.37891 200.021 8.78516V10.6484C200.021 10.9766 200.039 11.2637 200.074 11.5098C200.113 11.7559 200.168 11.9668 200.238 12.1426C200.312 12.3184 200.4 12.4629 200.502 12.5762C200.607 12.6855 200.729 12.7676 200.865 12.8223C201.002 12.873 201.152 12.8984 201.316 12.8984C201.523 12.8984 201.705 12.8574 201.861 12.7754C202.021 12.6934 202.156 12.5645 202.266 12.3887C202.375 12.209 202.457 11.9766 202.512 11.6914C202.57 11.4023 202.6 11.0547 202.6 10.6484ZM211.002 9.00195V10.4375C211.002 11.0938 210.936 11.6543 210.803 12.1191C210.67 12.584 210.479 12.9648 210.229 13.2617C209.979 13.5547 209.68 13.7715 209.332 13.9121C208.984 14.0488 208.598 14.1172 208.172 14.1172C207.832 14.1172 207.516 14.0742 207.223 13.9883C206.93 13.8984 206.666 13.7617 206.432 13.5781C206.201 13.3906 206.002 13.1523 205.834 12.8633C205.67 12.5742 205.543 12.2285 205.453 11.8262C205.367 11.4238 205.324 10.9609 205.324 10.4375V9.00195C205.324 8.3457 205.391 7.78711 205.523 7.32617C205.66 6.86523 205.854 6.48828 206.104 6.19531C206.354 5.90234 206.652 5.68945 207 5.55664C207.348 5.41992 207.734 5.35156 208.16 5.35156C208.504 5.35156 208.82 5.39453 209.109 5.48047C209.402 5.56641 209.664 5.70117 209.895 5.88477C210.129 6.06445 210.328 6.29883 210.492 6.58789C210.656 6.87305 210.781 7.2168 210.867 7.61914C210.957 8.01758 211.002 8.47852 211.002 9.00195ZM209.455 10.6484V8.78516C209.455 8.45703 209.436 8.17188 209.396 7.92969C209.361 7.68359 209.309 7.47461 209.238 7.30273C209.168 7.13086 209.08 6.99219 208.975 6.88672C208.869 6.77734 208.748 6.69727 208.611 6.64648C208.475 6.5957 208.324 6.57031 208.16 6.57031C207.957 6.57031 207.775 6.61133 207.615 6.69336C207.455 6.77148 207.32 6.89648 207.211 7.06836C207.102 7.24023 207.018 7.46875 206.959 7.75391C206.904 8.03516 206.877 8.37891 206.877 8.78516V10.6484C206.877 10.9766 206.895 11.2637 206.93 11.5098C206.969 11.7559 207.023 11.9668 207.094 12.1426C207.168 12.3184 207.256 12.4629 207.357 12.5762C207.463 12.6855 207.584 12.7676 207.721 12.8223C207.857 12.873 208.008 12.8984 208.172 12.8984C208.379 12.8984 208.561 12.8574 208.717 12.7754C208.877 12.6934 209.012 12.5645 209.121 12.3887C209.23 12.209 209.312 11.9766 209.367 11.6914C209.426 11.4023 209.455 11.0547 209.455 10.6484ZM218.789 6.76367L216.357 14H214.652L217.84 5.46875H218.93L218.789 6.76367ZM220.822 14L218.379 6.76367L218.232 5.46875H219.328L222.533 14H220.822ZM220.711 10.8301V12.1016H216.135V10.8301H220.711ZM224.115 5.46875H225.51L227.842 11.8906L230.168 5.46875H231.568L228.398 14H227.279L224.115 5.46875ZM223.412 5.46875H224.777L225.023 11.3691V14H223.412V5.46875ZM230.9 5.46875H232.271V14H230.66V11.3691L230.9 5.46875Z" fill="#111111" />
-      <path d="M350.752 5.42188V14H349.668V6.77539L347.482 7.57227V6.59375L350.582 5.42188H350.752ZM359.295 9.05469V10.3555C359.295 11.0547 359.232 11.6445 359.107 12.125C358.982 12.6055 358.803 12.9922 358.568 13.2852C358.334 13.5781 358.051 13.791 357.719 13.9238C357.391 14.0527 357.02 14.1172 356.605 14.1172C356.277 14.1172 355.975 14.0762 355.697 13.9941C355.42 13.9121 355.17 13.7812 354.947 13.6016C354.729 13.418 354.541 13.1797 354.385 12.8867C354.229 12.5938 354.109 12.2383 354.027 11.8203C353.945 11.4023 353.904 10.9141 353.904 10.3555V9.05469C353.904 8.35547 353.967 7.76953 354.092 7.29688C354.221 6.82422 354.402 6.44531 354.637 6.16016C354.871 5.87109 355.152 5.66406 355.48 5.53906C355.812 5.41406 356.184 5.35156 356.594 5.35156C356.926 5.35156 357.23 5.39258 357.508 5.47461C357.789 5.55273 358.039 5.67969 358.258 5.85547C358.477 6.02734 358.662 6.25781 358.814 6.54688C358.971 6.83203 359.09 7.18164 359.172 7.5957C359.254 8.00977 359.295 8.49609 359.295 9.05469ZM358.205 10.5312V8.87305C358.205 8.49023 358.182 8.1543 358.135 7.86523C358.092 7.57227 358.027 7.32227 357.941 7.11523C357.855 6.9082 357.746 6.74023 357.613 6.61133C357.484 6.48242 357.334 6.38867 357.162 6.33008C356.994 6.26758 356.805 6.23633 356.594 6.23633C356.336 6.23633 356.107 6.28516 355.908 6.38281C355.709 6.47656 355.541 6.62695 355.404 6.83398C355.271 7.04102 355.17 7.3125 355.1 7.64844C355.029 7.98438 354.994 8.39258 354.994 8.87305V10.5312C354.994 10.9141 355.016 11.252 355.059 11.5449C355.105 11.8379 355.174 12.0918 355.264 12.3066C355.354 12.5176 355.463 12.6914 355.592 12.8281C355.721 12.9648 355.869 13.0664 356.037 13.1328C356.209 13.1953 356.398 13.2266 356.605 13.2266C356.871 13.2266 357.104 13.1758 357.303 13.0742C357.502 12.9727 357.668 12.8145 357.801 12.5996C357.938 12.3809 358.039 12.1016 358.105 11.7617C358.172 11.418 358.205 11.0078 358.205 10.5312ZM366.045 9.05469V10.3555C366.045 11.0547 365.982 11.6445 365.857 12.125C365.732 12.6055 365.553 12.9922 365.318 13.2852C365.084 13.5781 364.801 13.791 364.469 13.9238C364.141 14.0527 363.77 14.1172 363.355 14.1172C363.027 14.1172 362.725 14.0762 362.447 13.9941C362.17 13.9121 361.92 13.7812 361.697 13.6016C361.479 13.418 361.291 13.1797 361.135 12.8867C360.979 12.5938 360.859 12.2383 360.777 11.8203C360.695 11.4023 360.654 10.9141 360.654 10.3555V9.05469C360.654 8.35547 360.717 7.76953 360.842 7.29688C360.971 6.82422 361.152 6.44531 361.387 6.16016C361.621 5.87109 361.902 5.66406 362.23 5.53906C362.562 5.41406 362.934 5.35156 363.344 5.35156C363.676 5.35156 363.98 5.39258 364.258 5.47461C364.539 5.55273 364.789 5.67969 365.008 5.85547C365.227 6.02734 365.412 6.25781 365.564 6.54688C365.721 6.83203 365.84 7.18164 365.922 7.5957C366.004 8.00977 366.045 8.49609 366.045 9.05469ZM364.955 10.5312V8.87305C364.955 8.49023 364.932 8.1543 364.885 7.86523C364.842 7.57227 364.777 7.32227 364.691 7.11523C364.605 6.9082 364.496 6.74023 364.363 6.61133C364.234 6.48242 364.084 6.38867 363.912 6.33008C363.744 6.26758 363.555 6.23633 363.344 6.23633C363.086 6.23633 362.857 6.28516 362.658 6.38281C362.459 6.47656 362.291 6.62695 362.154 6.83398C362.021 7.04102 361.92 7.3125 361.85 7.64844C361.779 7.98438 361.744 8.39258 361.744 8.87305V10.5312C361.744 10.9141 361.766 11.252 361.809 11.5449C361.855 11.8379 361.924 12.0918 362.014 12.3066C362.104 12.5176 362.213 12.6914 362.342 12.8281C362.471 12.9648 362.619 13.0664 362.787 13.1328C362.959 13.1953 363.148 13.2266 363.355 13.2266C363.621 13.2266 363.854 13.1758 364.053 13.0742C364.252 12.9727 364.418 12.8145 364.551 12.5996C364.688 12.3809 364.789 12.1016 364.855 11.7617C364.922 11.418 364.955 11.0078 364.955 10.5312ZM367.346 7.56055V7.10938C367.346 6.78516 367.416 6.49023 367.557 6.22461C367.697 5.95898 367.898 5.74609 368.16 5.58594C368.422 5.42578 368.732 5.3457 369.092 5.3457C369.459 5.3457 369.771 5.42578 370.029 5.58594C370.291 5.74609 370.492 5.95898 370.633 6.22461C370.773 6.49023 370.844 6.78516 370.844 7.10938V7.56055C370.844 7.87695 370.773 8.16797 370.633 8.43359C370.496 8.69922 370.297 8.91211 370.035 9.07227C369.777 9.23242 369.467 9.3125 369.104 9.3125C368.74 9.3125 368.426 9.23242 368.16 9.07227C367.898 8.91211 367.697 8.69922 367.557 8.43359C367.416 8.16797 367.346 7.87695 367.346 7.56055ZM368.16 7.10938V7.56055C368.16 7.74023 368.193 7.91016 368.26 8.07031C368.33 8.23047 368.436 8.36133 368.576 8.46289C368.717 8.56055 368.893 8.60938 369.104 8.60938C369.314 8.60938 369.488 8.56055 369.625 8.46289C369.762 8.36133 369.863 8.23047 369.93 8.07031C369.996 7.91016 370.029 7.74023 370.029 7.56055V7.10938C370.029 6.92578 369.994 6.75391 369.924 6.59375C369.857 6.42969 369.754 6.29883 369.613 6.20117C369.477 6.09961 369.303 6.04883 369.092 6.04883C368.885 6.04883 368.711 6.09961 368.57 6.20117C368.434 6.29883 368.33 6.42969 368.26 6.59375C368.193 6.75391 368.16 6.92578 368.16 7.10938ZM371.5 12.3652V11.9082C371.5 11.5879 371.57 11.2949 371.711 11.0293C371.852 10.7637 372.053 10.5508 372.314 10.3906C372.576 10.2305 372.887 10.1504 373.246 10.1504C373.613 10.1504 373.926 10.2305 374.184 10.3906C374.445 10.5508 374.646 10.7637 374.787 11.0293C374.928 11.2949 374.998 11.5879 374.998 11.9082V12.3652C374.998 12.6855 374.928 12.9785 374.787 13.2441C374.65 13.5098 374.451 13.7227 374.189 13.8828C373.932 14.043 373.621 14.123 373.258 14.123C372.895 14.123 372.582 14.043 372.32 13.8828C372.059 13.7227 371.855 13.5098 371.711 13.2441C371.57 12.9785 371.5 12.6855 371.5 12.3652ZM372.314 11.9082V12.3652C372.314 12.5449 372.348 12.7168 372.414 12.8809C372.484 13.041 372.59 13.1719 372.73 13.2734C372.871 13.3711 373.047 13.4199 373.258 13.4199C373.469 13.4199 373.643 13.3711 373.779 13.2734C373.92 13.1719 374.023 13.041 374.09 12.8809C374.156 12.7207 374.189 12.5488 374.189 12.3652V11.9082C374.189 11.7246 374.154 11.5527 374.084 11.3926C374.018 11.2324 373.914 11.1035 373.773 11.0059C373.637 10.9043 373.461 10.8535 373.246 10.8535C373.039 10.8535 372.865 10.9043 372.725 11.0059C372.588 11.1035 372.484 11.2324 372.414 11.3926C372.348 11.5527 372.314 11.7246 372.314 11.9082ZM373.516 6.68164L369.35 13.3496L368.74 12.9629L372.906 6.29492L373.516 6.68164Z" fill="#111111" />
-      <rect x="384.036" y="5.15387" width="21.3818" height="9.69231" rx="1.33" fill="#111111" />
-      <path opacity="0.4" fill-rule="evenodd" clip-rule="evenodd" d="M387.6 4H401.855C402.851 4 403.538 4.00078 404.072 4.04436C404.594 4.087 404.88 4.16555 405.09 4.27248C405.56 4.51217 405.942 4.89462 406.182 5.36502C406.289 5.5749 406.368 5.86085 406.41 6.38264C406.454 6.9161 406.455 7.60341 406.455 8.6V11.4C406.455 12.3966 406.454 13.0839 406.41 13.6174C406.368 14.1392 406.289 14.4251 406.182 14.635C405.942 15.1054 405.56 15.4878 405.09 15.7275C404.88 15.8345 404.594 15.913 404.072 15.9556C403.538 15.9992 402.851 16 401.855 16H387.6C386.603 16 385.916 15.9992 385.383 15.9556C384.861 15.913 384.575 15.8345 384.365 15.7275C383.895 15.4878 383.512 15.1054 383.272 14.635C383.166 14.4251 383.087 14.1392 383.044 13.6174C383.001 13.0839 383 12.3966 383 11.4V8.6C383 7.60341 383.001 6.9161 383.044 6.38264C383.087 5.86085 383.166 5.5749 383.272 5.36502C383.512 4.89462 383.895 4.51217 384.365 4.27248C384.575 4.16555 384.861 4.087 385.383 4.04436C385.916 4.00078 386.603 4 387.6 4ZM382 8.6C382 6.63982 382 5.65972 382.381 4.91103C382.717 4.25247 383.252 3.71703 383.911 3.38148C384.66 3 385.64 3 387.6 3H401.855C403.815 3 404.795 3 405.544 3.38148C406.202 3.71703 406.738 4.25247 407.073 4.91103C407.455 5.65972 407.455 6.63982 407.455 8.6V11.4C407.455 13.3602 407.455 14.3403 407.073 15.089C406.738 15.7475 406.202 16.283 405.544 16.6185C404.795 17 403.815 17 401.855 17H387.6C385.64 17 384.66 17 383.911 16.6185C383.252 16.283 382.717 15.7475 382.381 15.089C382 14.3403 382 13.3602 382 11.4V8.6ZM408.473 12.1538C409.316 11.9385 410 11.1895 410 10C410 8.81046 409.316 8.06154 408.473 7.84615L408.473 10L408.473 12.1538Z" fill="#111111" />
-      <defs>
-        <clipPath id="clip0_296_523">
-          <rect width="20" height="14" fill="white" transform="translate(4 3)" />
-        </clipPath>
-      </defs>
-    </svg>
-  )
-};
-
-export default StatusBar;
