--- src/frames/samsung-galaxy-s20-5g/svg/device-frame.tsx
@@ -1,18 +1,0 @@
-import { concat } from "nixix/primitives";
-import { deviceFrameHeightClass } from "~/constants";
-
-
-const DeviceFrame: Nixix.FC<App.SVGProps> = ({ className = ' ', ...rest }): someView => {
-  const resizeClass = ' tws-max-h-[971px] tws-w-auto tws-relative ';
-  return (
-    <svg {...rest} className={concat`${deviceFrameHeightClass} ${resizeClass} ${className} `} width="443" height="971" viewBox="0 0 443 971" fill="none" xmlns="http://www.w3.org/2000/svg">
-      <path d="M439.132 176.545H440.264C441.514 176.545 442.527 177.595 442.527 178.89V295.293C442.527 296.588 441.514 297.637 440.264 297.637H439.132V176.545Z" fill="#121515" />
-      <path d="M439.132 363.276H440.264C441.514 363.276 442.527 364.326 442.527 365.62V424.307C442.527 425.602 441.514 426.651 440.264 426.651H439.132V363.276Z" fill="#121515" />
-      <path d="M0 98.9107C0 54.8722 0 32.853 12.4737 18.2491C14.245 16.1754 16.1765 14.244 18.2504 12.4729C32.8553 0 54.8762 0 98.9178 0H340.214C384.256 0 406.276 0 420.881 12.4729C422.955 14.244 424.887 16.1754 426.658 18.2491C439.132 32.853 439.132 54.8722 439.132 98.9107V872.089C439.132 916.128 439.132 938.147 426.658 952.751C424.887 954.825 422.955 956.756 420.881 958.527C406.276 971 384.256 971 340.214 971H98.9178C54.8762 971 32.8553 971 18.2504 958.527C16.1765 956.756 14.245 954.825 12.4737 952.751C0 938.147 0 916.128 0 872.089V98.9107Z" fill="#3A4245" />
-      <path d="M2.26367 91.498C2.26367 54.4781 2.26367 35.9682 12.2256 23.3324C14.4845 20.4672 17.0735 17.8784 19.9389 15.6197C32.5756 5.65845 51.0868 5.65845 88.1094 5.65845H351.023C388.045 5.65845 406.556 5.65845 419.193 15.6197C422.058 17.8784 424.647 20.4672 426.906 23.3324C436.868 35.9682 436.868 54.4781 436.868 91.498V879.502C436.868 916.522 436.868 935.032 426.906 947.667C424.647 950.533 422.058 953.121 419.193 955.38C406.556 965.341 388.045 965.341 351.023 965.341H88.1094C51.0868 965.341 32.5756 965.341 19.9389 955.38C17.0735 953.121 14.4845 950.533 12.2256 947.667C2.26367 935.032 2.26367 916.522 2.26367 879.502V91.498Z" fill="#121515" />
-    </svg>
-  );
-};
-
-
-export default DeviceFrame;
