--- src/frames/samsung-galaxy-s20-5g/svg/virtual-home-button.tsx
@@ -1,14 +1,0 @@
-import { px } from "@/lib/utils";
-
-const VirtualHomeButton: Nixix.FC<JSX.IntrinsicElements['div']> = ({className = '', style, ...rest}): someView => {
-  return (
-    <div className={className} {...rest} style={{
-      height: px(4),
-      backgroundColor: 'black',
-      borderRadius: px(50),
-      ...style
-    }} />
-  )
-}
-
-export default VirtualHomeButton;
